% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_marking_data.R
\name{bootstrap_marking_data}
\alias{bootstrap_marking_data}
\title{This function bootstraps the marking data for every area of origin.}
\usage{
bootstrap_marking_data(mark_recapture_object)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}
}
\value{
mark_recapture_object with estimates of all selected parameters
}
\description{
This function bootstraps the marking data for every area of origin.
}
\examples{
mro <- bootstrap_marking_data(mro1D)

}
