% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapQuantiles.R
\name{bootstrapQuantiles}
\alias{bootstrapQuantiles}
\title{linear model of kernel density estimate}
\usage{
bootstrapQuantiles(markRecaptureObject, param = c("s", "m", "r"), filename)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{filename}{(path to file) and filename to store the markRecaptureObject
with bootstrap estimates.}

\item{parameters}{vector of characters specifying the parameter names
considered for uncertainty estimation, possible values are s for survival,
m for migratory connectivity and r for recovery probability. Defaults to
c("s","m","r").}
}
\value{
raw bootstrap dataframe to be used in bootstrapQuantiles
}
\description{
This function estimates the uncertainty of the parameters survival, migratory
connectivity and recovery probability by bootstrapping the marking data.
}
\examples{
mro <- bootstrapQuantiles(mro1D,"s")
}
