% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_discrete_m.R
\name{calc_discrete_m}
\alias{calc_discrete_m}
\title{True discrete migratory connectivity from a
continuous migratory connectivity function.}
\usage{
calc_discrete_m(mark_recapture_object, knots_prop)
}
\arguments{
\item{mark_recapture_object}{an object of the class 'mark_recapture_object'}

\item{knots_prop}{knot sequence, list with vectors for latitude and
longitude, each vector contains W+1 entries which specify the boundaries
between the discrete non-breeding areas}
}
\value{
mark_recapture_object
}
\description{
This function integrates the continuous migratory connectivity function
for disjoint intervals given as a knot sequence and adds the discrete
migratory connectivity values to the
mark_recapture_object$breedingAreas[[breedingAreaName]]$mDiscrete.
This function is for simulated data only.
}
\examples{
{
    kp <-list(latitude = c(0,0.25, 0.5, 0.75,1),
        longitude = NULL)
    mro <- calcDiscreteM(mark_recapture_object = mro1D, knots_prop = kp)
}
}
