% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiscreteM.R
\name{calcDiscreteM}
\alias{calcDiscreteM}
\title{True discrete migratory connectivity from a
continuous migratory connectivity function.}
\usage{
calcDiscreteM(markRecaptureObject, knots_prop)
}
\arguments{
\item{markRecaptureObject}{an object of the class 'markRecaptureObject'}

\item{knots_prop}{knot sequence, list with vectors for latitude and longitude,
each vector contains W+1 entries which specify the boundaries between the
discrete non-breeding areas}
}
\value{
markRecaptureObject
}
\description{
This function integrates the continuous migratory connectivity function
for disjoint intervals given as a knot sequence and adds the discrete
migratory connectivity values to the
markRecaptureObject$breedingAreas[[breedingAreaName]]$mDiscrete.
This function is for simulated data only.
}
\examples{
{
    kp <-list(latitude = c(0,0.25, 0.5, 0.75,1),
        longitude = NULL)
    mro <- calcDiscreteM(markRecaptureObject = mro1D, knots_prop = kp)
}
}
