% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_estimate.R
\name{comb_estimate}
\alias{comb_estimate}
\title{Function to get combined migratory connectivity estimate.}
\usage{
comb_estimate(
  optimization_object,
  start_times = 1,
  maxit = 1e+05,
  reltol = 1e-08,
  change_r = FALSE
)
}
\arguments{
\item{optimization_object}{an object of the class 'optimization_object' (see
constructors)}

\item{start_times}{number of repetitions of the optimization procedure.
Defaults to 1.}

\item{maxit}{numbers of iterations used by the control argument of optim,
defaults to 10000, see ?optim for details.}

\item{reltol}{Relative convergence tolerance to stop optimization algorithm.
Used by the control argument of optim,
defaults to 1e-8, see ?optim for details.}

\item{change_r}{logical. If TRUE, it estimates the combined version of the
recovery probability and adds it to
mark_recapture_object$estimates$r_combined.
Should be commonly set to TRUE if the breeding area argument in the
optimization Object is 'all' and FALSE otherwise. Defaults to FALSE.}
}
\value{
f_x data.frame with estimates of every optimization

val numeric vector of all optimizations

mark_recapture_object
}
\description{
This function uses an optimization procedure to correct the continuous
migratory connectivity estimator by the discrete migratory connectivity
estimator. Throws an error if convergence is not achieved.
}
\examples{
{
    data("mro1D_increasing")
    o_o <- optimization_object(mark_recapture_object = mro1D_increasing$mro,
        b = "all",
        split = mro1D_increasing$split,
        lambda  = c(.05,300))

    mro <- comb_estimate(optimization_object = o_o)
}
}
