% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combEstimate.R
\name{combEstimate}
\alias{combEstimate}
\title{Function to get combined migratory connectivity estimate.}
\usage{
combEstimate(
  optimizationObject,
  markRecaptureObject,
  startTimes = 1,
  maxit = 1e+05,
  reltol = 1e-08,
  changeR = FALSE
)
}
\arguments{
\item{optimizationObject}{an object of the class 'optimizationObject' (s. constructors)}

\item{markRecaptureObject}{an object of the class 'markRecaptureObject'}

\item{startTimes}{number of repetitions of the optimization procedure. Defaults to 1.}

\item{maxit}{numbers of iterations used by the control argument of optim,
defaults to 10000, see ?optim for details.}

\item{reltol}{Relative convergence tolerance to stop optimization algorithm.
Used by the control argument of optim,
defaults to 1e-8, see ?optim for details.}

\item{changeR}{logical. If TRUE, it estimates the combined version of the recovery
probability and adds it to markRecaptureObject$estimates$rCombined. Should be commonly
set to TRUE if the breeding area argument in the optimization Object is 'all' and FALSE
otherwise. Defaults to FALSE.}
}
\value{
fVonX data.frame with estimates of every optimization

val numeric vector of all optimizations

markRecaptureObject
}
\description{
This function uses an optimization procedure to correct the continuous migratory
connectivity estimator by the discrete migratory connectivity estimator. Throws an error
if convergence is not achieved.
}
\examples{
combEstimate()
}
