% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_owin_from_shp.R
\name{create_owin_from_shp}
\alias{create_owin_from_shp}
\title{creates an owin-object from a shpfile}
\usage{
create_owin_from_shp(file, crs, destination_field = NULL)
}
\arguments{
\item{file}{character, containing the path to the shape file. The shape file
includes at least the boundaries of the continuous non-breeding area. It may
contain fields not relevant for the analysis. For the combined approach one
field must contain information on to which discrete non-breeding area each
geometry belongs (see destination_field).}

\item{crs}{coordinate reference system.}

\item{destination_field}{only used for the combined model approach. Name of
the field containing the information on the discrete non-breeding area in
the shape file. Defaults to NULL.}
}
\value{
owin object only containing the geometry and, in case it is used for
the combined approach, information on the discrete non-breeding area each
geometry belongs to
}
\description{
This function creates an owin-object from a shpfile adding discrete
non-breeding areas if applicable.
}
\examples{
\dontrun{create_owin_from_shp(file, crs = "ESRI:54009")}
}
