% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOwinFromShp.R
\name{createOwinFromShp}
\alias{createOwinFromShp}
\title{creates an owin-object from a shpfile}
\usage{
createOwinFromShp(
  file,
  crs = "+proj=utm +zone=31N +datum=WGS84",
  nonbreedingField = NULL
)
}
\arguments{
\item{file}{character, containing the path to the shape file. The shape file
includes at least the boundaries of the continuous non-breeding area. It may
contain fields not relevant for the analysis. For the combined approach one
field must contain information on to which discrete non-breeding area each
geometry belongs (see nonbreedingField).}

\item{crs}{coordinate reference system. Defaults to
"+proj=utm +zone=31N +datum=WGS84". For details see ?sp::CRS.}

\item{nonbreedingField}{only used for the combined model approach. Name of
the field containing the information on the discrete non-breeding area in
the shape file. Defaults to NULL.}
}
\value{
owin object only containing the geometry and, in case it is used for
the combined approach, information on the discrete non-breeding area each
geometry belongs to
}
\description{
This function creates an owin-object from a shpfile adding discrete
non-breeding areas if applicable.
}
\examples{
\dontrun{createOwinFromShp(file)}
}
