% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_bspline.R
\name{define_bspline}
\alias{define_bspline}
\title{function to define a positive a b-spline function depending on its parameters}
\usage{
define_bspline(raw_spline, beta, inside)
}
\arguments{
\item{raw_spline}{raw spline, value of init_spline()}

\item{beta}{parameter of the b-spline, values are to be optimized}

\item{inside}{specifies if a cell of the gridded window is inside the window
of the data or not. Vector of logicals.}
}
\value{
function defining a b-spline for a given parameter vector
}
\description{
This function defines the exponential of a b-spline dependent on its
parameters beta
}
\examples{
{
    y <- seq(0,1,length.out=100)
    i_k <- seq(0.1111111,0.8888889,length.out=8)
    r_s <- init_spline(y = y,
        knots = i_k,
        degree = 3,
        intercept = TRUE,
        dim = 1)
    d_b <- define_bspline(raw_spline = r_s,
        beta,
        inside = rep(TRUE, 100))
    d_b(beta = rnorm(12))
}
}
