% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineBspline.R
\name{defineBspline}
\alias{defineBspline}
\title{function to define a positive a b-spline function depending on its parameters}
\usage{
defineBspline(rawSpline, beta, inside)
}
\arguments{
\item{rawSpline}{raw spline, value of initSpline()}

\item{beta}{parameter of the b-spline, values are to be optimized}

\item{inside}{specifies if a cell of the gridded window is inside the window of the data
or not. Vector of logicals.}
}
\value{
function defining a b-spline for a given parameter vector
}
\description{
This function defines the exponential of a b-spline dependent on its parameters beta
}
\examples{
defineBspline(w,beta,knots)
}
