% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{destination}
\alias{destination}
\title{helper function for destination area}
\usage{
destination(
  window = NULL,
  survival,
  recovery,
  xrange = c(0, 0),
  yrange = c(0, 0),
  crs,
  recovery_data = NULL
)
}
\arguments{
\item{window}{object of class "owin":observation window in
two-dimensional plane}

\item{survival}{function: survival function defined over the whole
destination area, independent of the origin}

\item{recovery}{constant function: recovery probability, must be
constant over whole destination area}

\item{xrange}{vector in the form of c(xmin,xmax). To define line or rectangle
xrange and yrange can be used instead of window.}

\item{yrange}{vector in the form of c(ymin,ymax). To define line or rectangle
xrange and yrange can be used instead of window.}

\item{crs}{coordinate system of destination}

\item{recovery_data}{empty space for simulated or real-world recovery data}
}
\value{
object of class "destination": contains list of window, survival
and recovery for the destination area
}
\description{
This function defines the properties of the destination area using the
constructor.
}
\examples{
{
     w_a <- destination(survival = function(w) 0.3,
     recovery = function(w) 0.01,
     xrange = c(0,1),
     crs = "EPSG:4326" )
 }
}
