% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lm.R
\name{est_lm}
\alias{est_lm}
\title{linear model of kernel density estimate}
\usage{
est_lm(mark_recapture_object, b, fixed_slope = NULL)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{b}{name of areas of origin}

\item{fixed_slope}{numeric. Value for the fixed slope, e.g., to estimate
a linear model for the areas of origin separately.}
}
\value{
vector of length res with survival probabilities dependent on space
}
\description{
This function estimates the parameters of a linear model for a certain point
of the kernel density estimate dependent on time. Linear models are obtained
according to the resolution.
}
\examples{
mro <- est_lm(mro1D, b = "all")
}
