% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_m.R
\name{est_m}
\alias{est_m}
\title{estimate migratory connectivity}
\usage{
est_m(mark_recapture_object, all = FALSE)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{all}{boolean: if TRUE only one kernel density estimate will be
calculated summarizing all areas of origin. Defaults to FALSE.}
}
\value{
list of vectors with length res-1 containing migratory connectivity
density of every spot
}
\description{
This function estimates the migratory connectivity in space constant over
time when survival and raw distribution of dead recoveries for every
area of origin is known.
}
\examples{
mro <- est_m(mro1D, all = TRUE)
}
