% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_parameters.R
\name{est_parameters}
\alias{est_parameters}
\title{linear model of kernel density estimate}
\usage{
est_parameters(
  mark_recapture_object,
  res = NULL,
  parameters = c("s", "m", "r")
)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{parameters}{vector of characters specifying the parameter names
considered for uncertainty estimation, possible values are s for survival,
m for migratory connectivity and r for recovery probability. Defaults to
c("s","m","r").}
}
\value{
mark_recapture_object with estimates of all selected parameters
}
\description{
This function estimates the uncertainty of the parameters survival, migratory
connectivity and recovery probability by bootstrapping the marking data.
}
\examples{
mro <- est_parameters(mro1D)

}
