% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_r.R
\name{est_r}
\alias{est_r}
\title{recovery probability estimator}
\usage{
est_r(mark_recapture_object, b = "all")
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{b}{name of the area of origin to estimate recovery for. Defaults to
"all".}
}
\value{
mark_recpature_object with scalar of recovery probability
}
\description{
This function estimates the recovery probability uniformly over the whole
recovery area.
}
\examples{
mro <- est_r(mro1D)
}
