% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_s.R
\name{est_s}
\alias{est_s}
\title{survival estimator}
\usage{
est_s(mark_recapture_object)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}
}
\value{
mark_recapture_object vector/matrix of length res with survival
probabilities dependent on space
}
\description{
This function estimates the survival from a kernel density estimate of the
data of recovered individuals. It uses the data of all areas of origin at
once.
}
\examples{
mro <- est_s(mro1D)
}
