% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_uncertainty.R
\name{est_uncertainty}
\alias{est_uncertainty}
\title{linear model of kernel density estimate}
\usage{
est_uncertainty(
  mark_recapture_object,
  parameters = c("s", "m", "r"),
  iterations = NULL,
  bootstrap_data = NULL,
  res = NULL,
  filename
)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{parameters}{vector of characters specifying the parameter names
considered for uncertainty estimation, possible values are s for survival,
m for migratory connectivity and r for recovery probability. Defaults to
c("s","m","r").}

\item{iterations}{numeric. Number of bootstraps performed. Defaults to NULL.}

\item{bootstrap_data}{data prepared by bootstrapMarkingData, offers the
option to compute bootstrap in parallel. Defauls to NULL.}

\item{res}{spatial resolution. Defaults to NULL. If NULL, resolution of the
mark_recapture_object$spatial_resolution is used.}

\item{filename}{(path to file) and filename to store the
mark_recapture_object with bootstrap estimates.}
}
\value{
mark_recapture_object with added bootstrap uncertainty for
parameters
}
\description{
This function estimates the uncertainty of the parameters survival, migratory
connectivity and recovery probability by bootstrapping the marking data.
}
\examples{
{
mro  <- mro1D
mro <- est_uncertainty(mro, parameters = "s",
iterations = 1, filename = "test")
}

}
