% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estLM.R
\name{estLM}
\alias{estLM}
\title{linear model of kernel density estimate}
\usage{
estLM(res_x, markRecaptureObject, res_y = res_x, b, dataType = "sim",
  robust = T, auxiliaryVariable = NULL)
}
\arguments{
\item{res_x}{resolution in space}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}
}
\value{
vector of length res_x with survival probabilities dependent on space
}
\description{
This function estimates the parameters of a linear model for a certain point
of the kernel density estimate dependent on time. Linear models are obtained
according to the resolution.
}
\examples{
estS()
}
