% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estParameters.R
\name{estParameters}
\alias{estParameters}
\title{linear model of kernel density estimate}
\usage{
estParameters(markRecaptureObject, parameters = c("s", "m", "r"))
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{parameters}{vector of characters specifying the parameter names
considered for uncertainty estimation, possible values are s for survival,
m for migratory connectivity and r for recovery probability. Defaults to
c("s","m","r").}
}
\value{
markRecaptureObject with estimates of all selected parameters
}
\description{
This function estimates the uncertainty of the parameters survival, migratory
connectivity and recovery probability by bootstrapping the marking data.
}
\examples{
mro <- estParameters(mro1D)
}
