% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estR.R
\name{estR}
\alias{estR}
\title{recovery probability estimator}
\usage{
estR(markRecaptureObject, b = "all")
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{b}{name of breeding area to estimate recovery for. Defaults to "all".}
}
\value{
scalar of recovery probability
}
\description{
This function estimates the recovery probability uniformely over the whole recovery area.
}
\examples{
estR()

}
