% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estUncertainty.R
\name{estUncertainty}
\alias{estUncertainty}
\title{linear model of kernel density estimate}
\usage{
estUncertainty(
  markRecaptureObject,
  parameters = c("s", "m", "r"),
  iterations = NULL,
  bootstrapData = NULL,
  res = NULL,
  filename
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{parameters}{vector of characters specifying the parameter names
considered for uncertainty estimation, possible values are s for survival,
m for migratory connectivity and r for recovery probability. Defaults to
c("s","m","r").}

\item{iterations}{numeric. Number of bootstraps performed. Defaults to NULL.}

\item{bootstrapData}{data prepared by bootstrapMarkingData, offers the
option to compute bootstrap in parallel. Defauls to NULL.}

\item{res}{spatial resolution. Defaults to NULL. If NULL, resolution of the
markRecaptureObject$spatialResolution is used.}

\item{filename}{(path to file) and filename to store the markRecaptureObject
with bootstrap estimates.}
}
\value{
markRecaptureObject with added bootstrap uncertainty for parameters
}
\description{
This function estimates the uncertainty of the parameters survival, migratory
connectivity and recovery probability by bootstrapping the marking data.
}
\examples{
mro <- estUncertainty(mro1D)
}
