% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bootstrap_parameters.R
\name{get_bootstrap_parameters}
\alias{get_bootstrap_parameters}
\title{linear model of kernel density estimate}
\usage{
get_bootstrap_parameters(mark_recapture_object, param)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{parameter}{vector of characters specifying the parameter names
considered for uncertainty estimation, possible values are s for survival,
m for migratory connectivity and r for recovery probability. Defaults to
c("s","m","r").}
}
\value{
raw bootstrap dataframe to be used in bootstrap_quantiles
}
\description{
This function estimates the uncertainty of the parameters survival, migratory
connectivity and recovery probability by bootstrapping the marking data.
}
\examples{
mro <- get_bootstrap_parameters(mro1D, "s")

}
