% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBootstrapParameters.R
\name{getBootstrapParameters}
\alias{getBootstrapParameters}
\title{linear model of kernel density estimate}
\usage{
getBootstrapParameters(markRecaptureObject, param)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{parameter}{vector of characters specifying the parameter names
considered for uncertainty estimation, possible values are s for survival,
m for migratory connectivity and r for recovery probability. Defaults to
c("s","m","r").}
}
\value{
raw bootstrap dataframe to be used in bootstrapQuantiles
}
\description{
This function estimates the uncertainty of the parameters survival, migratory
connectivity and recovery probability by bootstrapping the marking data.
}
\examples{
mro <- getBootstrapParameters(mro1D,"s")
}
