% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_integrate_dist_continuous.R
\name{gr_integrate_dist_continuous}
\alias{gr_integrate_dist_continuous}
\title{gradient of function to define the distance between a b-spline representing a
density and the continuous migratory connectivity estimate}
\usage{
gr_integrate_dist_continuous(beta, k, raw_spline, dim, m, inside, normalize)
}
\arguments{
\item{beta}{parameter of the b-spline, values are to be optimized}

\item{k}{number of parameters}

\item{raw_spline}{raw spline, value of init_spline()}

\item{dim}{spatial dimension of the data}

\item{m}{vector of continuous migratory connectivity estimate}

\item{inside}{specifies if a cell of the gridded window is inside the window
of the data or not. Vector of logicals.}

\item{normalize}{numeric, normalizes the discretized integralt. Equals to the
spatial resolution in one-dimensional space and to the
product of the spatial resolutions in two-dimensional space.}
}
\value{
function defining the distance between a b-spline and continuous
        migratory connectivity
}
\description{
This function numerically integrates the quadratic distance between a
b-spline representing a density and the continuous migratory connectivity
estimate.
}
\examples{
{
    y <- seq(0,1,length.out=100)
    i_k <- seq(0.1111111,0.8888889,length.out=8)
    r_s <- initSpline(y = y,
        knots = i_k,
        degree = 3,
        intercept = TRUE,
        dim = 1)
    gr_ic <- gr_integrate_dist_continuous(beta, k,
        raw_spline = r_s,
        dim = 1,
        m = mro1D_increasing$mro$estimates$m$all,
        inside = rep(1, 100),
        normalize = 100)
   gr_ic(beta = rnorm(12), k = 2)
}
}
