% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_lh.R
\name{gr_lh}
\alias{gr_lh}
\title{gradient of function defining the smoothness of a bspline}
\usage{
gr_lh(beta, k, dim, A, normalize)
}
\arguments{
\item{beta}{vector of parameters of the b-spline}

\item{k}{number of parameters}

\item{dim}{dimension of space}

\item{A}{matrix of second derivative of b-spline}

\item{normalize}{numeric, normalizes the discretized integralt. Equals to the
spatial resolution in one-dimensional space and to the product of the spatial
resolutions in two-dimensional space.}
}
\value{
function defining the distance between a b-spline and discrete
        migratory connectivity depending on the parameters
}
\description{
This function defines the smoothness of a bspline as the second derivative
}
\examples{
{
    y <- seq(0, 1, length.out=100)
    i_k <- seq(0.1111111, 0.8888889, length.out=8)
    gr_l <- gr_lh(beta,k,dim = 1,
        A = splines2::dbs(y, knots = i_k, derivs = 2, degree = 3,
                          intercept = TRUE),
        normalize = 100)
    gr_l(beta = rnorm(12), k = 2)
}
}
