% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_bootstrapped_datasets.R
\name{init_bootstrapped_datasets}
\alias{init_bootstrapped_datasets}
\title{This function initializes a list of bootstrapped data sets for which in a
next step the parameters may be estimated.}
\usage{
init_bootstrapped_datasets(iterations = 1000, mark_recapture_object, filename)
}
\arguments{
\item{iterations}{number of bootstraps}

\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{filename}{(path to file) and filename to store list of bootstrapped
data.}
}
\value{
list of bootstrapped datasets
}
\description{
This function initializes a list of bootstrapped data sets for which in a
next step the parameters may be estimated.
}
\examples{
bootstrap_mro <- init_bootstrapped_datasets(
  mark_recapture_object = mro1D,
  filename = "mro1D_bootstrap_raw"
)

}
