% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeBootstrappedDatasets.R
\name{initializeBootstrappedDatasets}
\alias{initializeBootstrappedDatasets}
\title{This function initializes a list of bootstrapped data sets for which in a
next step the parameters may be estimated.}
\usage{
initializeBootstrappedDatasets(
  iterations = 1000,
  markRecaptureObject,
  filename
)
}
\arguments{
\item{iterations}{number of bootstraps}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{filename}{(path to file) and filename to store list of bootstrapped
data.}
}
\value{
list of bootstrapped datasets
}
\description{
This function initializes a list of bootstrapped data sets for which in a
next step the parameters may be estimated.
}
\examples{
bootstrap_mro <- initializeBootstrappedDatasets(
                                markRecaptureObject = mro1D,
                                filename = "mro1D_bootstrap_raw")
}
