% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSpline.R
\name{initSpline}
\alias{initSpline}
\title{Function to initialize a b-spline for optimization.}
\usage{
initSpline(y, knots, degree, intercept = TRUE, dim)
}
\arguments{
\item{y}{numeric vector, helper sequence, has to span whole wintering area,
defaults to seq(0,1,length.out = 100)}

\item{knots}{numeric vector, inner knot sequence to span the b-spline}

\item{degree}{integer, degree of the piecewise polynomial the b-spline is
made of, defaults to 3 (cubic b-spline)}

\item{intercept}{logical, defines if intercept is estimated or not, defaults
to TRUE, do not change default without good reason!}

\item{dim}{spatial dimension. Integer. Can be 1 or 2.}
}
\value{
returns a bSpline. See ?splines2::bSpline for details.
}
\description{
This function initilizes a b-spline on a helper sequence to be used for
optimization.
}
\examples{
{
    y <- seq(0,1,length.out=100)
    iK <- seq(0.1111111,0.8888889,length.out=8)
    iS <- initSpline(y = y, knots = iK, degree = 3, dim = 1)
}
}
