% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_dist_discrete.R
\name{integrate_dist_discrete}
\alias{integrate_dist_discrete}
\title{function to define the distance between a b-spline representing a density and
the discrete proportion}
\usage{
integrate_dist_discrete(
  raw_spline,
  dim,
  split,
  beta,
  b,
  prop,
  print = TRUE,
  inside
)
}
\arguments{
\item{raw_spline}{raw spline, value of init_spline()}

\item{dim}{spatial dimension of the data}

\item{split}{vector of length of y which defines the affiliation to a
discrete wintering area}

\item{beta}{parameter of the b-spline, values are to be optimized}

\item{b}{name of breeding area}

\item{prop}{vector of proportions of individuals going to discrete wintering
areas (discrete estimate or expected value for migratory
connectivity)}

\item{print}{logical, should proportions be printed or not?}

\item{inside}{specifies if a cell of the gridded window is inside the window
of the data or not. Vector of logicals.}
}
\value{
function defining the distance between a b-spline and discrete
        migratory connectivity depending on the parameters
}
\description{
This function numerically integrates the quadratic distance between a
b-spline representing a density and the discrete migratory connectivity
}
\examples{
{
    y <- seq(0, 1, length.out=100)
    i_k <- seq(0.1111111, 0.8888889, length.out = 8)
    r_s <- init_spline(y = y,
        knots = i_k,
        degree = 3,
        intercept = TRUE,
        dim = 1)
    i_d <- integrate_dist_discrete(raw_spline = r_s, dim = 1,
        split = mro1D_increasing$split, beta, b = "all",
        prop = mro1D_increasing$mro$origins$all$m_discrete /
            sum(mro1D_increasing$mro$origins$all$m_discrete),
        inside = rep(TRUE, 100))
    i_d(rnorm(12))
}
}
