% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateDist2Continuous.R
\name{integrateDist2Continuous}
\alias{integrateDist2Continuous}
\title{function to define the distance between a b-spline representing a density
and the continuous migratory connectivity estimate}
\usage{
integrateDist2Continuous(rawSpline, dim, beta, m, res, inside, normalize)
}
\arguments{
\item{rawSpline}{raw spline, value of initSpline()}

\item{dim}{spatial dimension of the data}

\item{beta}{parameter of the b-spline, values are to be optimized}

\item{m}{vector of continuous migratory connectivity estimate}

\item{res}{spatial resolution of estimates}

\item{inside}{specifies if a cell of the gridded window is inside the window of the data
or not. Vector of logicals.}
}
\value{
function defining the distance between a b-spline and
        continuous migratory connectivity
}
\description{
This function numerically integrates the quadratic distance between a b-spline
representing a density and the continuous migratory connectivity estimate
}
\examples{
integrateDist2Continuous()
}
