% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateDist2Discrete.R
\name{integrateDist2Discrete}
\alias{integrateDist2Discrete}
\title{function to define the distance between a b-spline representing a density and
the discrete proportion}
\usage{
integrateDist2Discrete(
  rawSpline,
  dim,
  split,
  beta,
  b,
  prop,
  print = TRUE,
  inside
)
}
\arguments{
\item{rawSpline}{raw spline, value of initSpline()}

\item{dim}{spatial dimension of the data}

\item{split}{vector of length of y which defines the affiliation to a
discrete wintering area}

\item{beta}{parameter of the b-spline, values are to be optimized}

\item{b}{name of breeding area}

\item{prop}{vector of proportions of individuals going to discrete wintering
areas (discrete estimate or expected value for migratory
connectivity)}

\item{print}{logical, should proportions be printed or not?}

\item{inside}{specifies if a cell of the gridded window is inside the window
of the data or not. Vector of logicals.}
}
\value{
function defining the distance between a b-spline and discrete
        migratory connectivity depending on the parameters
}
\description{
This function numerically integrates the quadratic distance between a
b-spline representing a density and the discrete migratory connectivity
}
\examples{
{
    y <- seq(0,1,length.out=100)
    iK <- seq(0.1111111,0.8888889,length.out=8)
    rS <- initSpline(y=y,
        knots = iK,
        degree = 3,
        intercept = TRUE,
        dim = 1)
    iD <- integrateDist2Discrete(rawSpline = rS, dim = 1,
        split = mro1DIncreasing$split, beta, b = "all",
        prop = mro1DIncreasing$mro$breedingAreas$all$mDiscrete /
            sum(mro1DIncreasing$mro$breedingAreas$all$mDiscrete),
        inside = rep(TRUE,100))
    iD(rnorm(12))
}
}
