% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lh.R
\name{Lh}
\alias{Lh}
\title{function defining the smoothness of a bspline}
\usage{
Lh(beta, dim, A, normalize)
}
\arguments{
\item{beta}{vector of parameters of the b-spline}

\item{dim}{dimension of space}

\item{A}{matrix of second derivative of b-spline}

\item{normalize}{factor to normalize with.}
}
\value{
function defining the distance between a b-spline and
        discrete migratory connectivity depending on the parameters
}
\description{
This function defines the smoothness of a bspline as the second derivative
}
\examples{
{
    y <- seq(0,1,length.out=100)
    iK <- seq(0.1111111,0.8888889,length.out=8)
     A <- splines2::dbs(y,knots=iK,derivs = 2, degree = 3, intercept = TRUE)
    L <- Lh(beta,dim = 1,
        A = t(A)\%*\%A,
        normalize = 100)
    L(beta = rnorm(12))
}
}
