% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{mark_recapture_object}
\alias{mark_recapture_object}
\title{helper function for mark recapture object}
\usage{
mark_recapture_object(
  window = NULL,
  xrange = c(0, 0),
  yrange = c(0, 0),
  survival = NULL,
  recovery = NULL,
  marked_individuals,
  migratory_connectivity = NULL,
  observation_time,
  real_recoveries = NULL,
  origin_names = NULL,
  robust = TRUE,
  crs = "EPSG:4326"
)
}
\arguments{
\item{window}{object of class "owin":observation window in
two-dimensional plane}

\item{xrange}{vector in the form of c(xmin,xmax). To define line or rectangle
xrange and yrange can be used instead of window.}

\item{yrange}{vector in the form of c(ymin,ymax). To define line or rectangle
xrange and yrange can be used instead of window.}

\item{survival}{function: survival function defined over the whole
destination area, independent of the origin}

\item{recovery}{constant function: recovery probability, must be
constant over whole destination area}

\item{marked_individuals}{integer: number of marked individuals in this
origin}

\item{migratory_connectivity}{either a list of functions containing one
migratory connectivity function for every origin or a function with
parameter b, allowing to partialize the function for every origin with
purrr::partial}

\item{observation_time}{length of observation window in years}

\item{real_recoveries}{real-world recovery data, defaults to NULL}

\item{origin_names}{character vector with origin names, defaults
to NULL}

\item{robust}{logical if TRUE robust linear model is calculated to estimate
survival and migratory connectivity}

\item{crs}{coordinate system. Defaults to "EPSG:4326" (longitude/latitude).}
}
\value{
object of class "mark_recapture_object": contains a list of the
destination, origins, observation time, number of origins and
spatial dimension
}
\description{
This function defines the properties of the mark recapture object using the
constructor.
}
\examples{
{
mro <- mark_recapture_object(xrange = c(0,1),
    survival = function(w) {0.5*w+.4},
    recovery = function(w) {0.01},
    marked_individuals = rep(100000,5) ,
    migratory_connectivity = function(b,w,B=B) {
        truncnorm::dtruncnorm(w,0,1, mean = seq(0.1,0.9,length.out = B)[b],
        sd = 0.3)
        },
    observation_time = 10)
}
}
