% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mro1D.R
\docType{data}
\name{mro1D}
\alias{mro1D}
\title{simulated data for the continuous approach (1D)}
\format{
A markRecaptureObject.
\describe{
  \item{mro1D}{An object of the type mark_recapture_object containing the
  simulated data with the true values of the parameters and the estimated
  values of the continuous approach}
}
}
\usage{
data(mro1D)
}
\description{
A data set containing data simulated with constant recovery probability in
one-dimensional space and continuous estimates of all parameters.
}
\examples{

data(mro1D)
str(mro1D)
}
\references{
This data set was artificially created for the CONSURE package.
}
\keyword{data}
\keyword{sets}
