% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mro1D_increasing.R
\docType{data}
\name{mro1D_increasing}
\alias{mro1D_increasing}
\title{simulated data for the combined approach (1D)}
\format{
A mark_recapture_object
\describe{
  \item{mro}{An object of the type mark_recapture_object containing
  the simulated data with the true values of the parameters and the
  estimated values of the continuous approach}
  \item{split}{numeric vector indicating for each point in space to which
   discrete wintering area it belongs}
}
}
\usage{
data(mro1D_increasing)
}
\description{
A dataset containing data simulated with increasing recovery probability in
one-dimensional space and continuous and combined estimates of all
parameters.
}
\examples{

data(mro1D_increasing)
str(mro1D_increasing)
}
\references{
This data set was artificially created for the CONSURE package.
}
\keyword{datasets}
