% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mro1D_increasing.R
\docType{data}
\name{mro1DIncreasing}
\alias{mro1DIncreasing}
\title{simulated data for the combined approach (1D)}
\format{
A markRecaptureObject.
\describe{
  \item{mro1DIncreasing}{An object of the type markRecaptureObject containing
  the simulated data with the true values of the parameteres and the
  estimated values of the continuous approach}
  \item{split}{numeric vector indicating for each point in space to which
   discrete wintering area it belongs}
}
}
\usage{
data(mro1DIncreasing)
}
\description{
A dataset containing data simulated with increasing recovery probability in
one-dimensional space and continuous and combined estimates of all
parameters.
}
\examples{

data(mro1DIncreasing)
str(mro1DIncreasing)
}
\references{
This data set was artificially created for the CONSURE package.
}
\keyword{datasets}
