% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_destination}
\alias{new_destination}
\title{constructor for destination area}
\usage{
new_destination(
  window = spatstat.geom::owin(),
  crs,
  survival,
  recovery,
  recovery_data
)
}
\arguments{
\item{window}{object of class "owin":observation window in
two-dimensional plane}

\item{survival}{function: survival function defined over the whole
destination area, independent of the origin}

\item{recovery}{constant function: recovery probability, must be
constant over whole destination area}

\item{recovery_data}{empty space for simulated or real-world recovery data}
}
\value{
object of class "destination": contains list of window, survival
        and recovery for the destination area
}
\description{
This function defines the properties of the destination area.
}
