% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_mark_recapture_object}
\alias{new_mark_recapture_object}
\title{constructor for mark recapture object}
\usage{
new_mark_recapture_object(
  destination,
  origins,
  observation_time,
  number_of_origins,
  spatial_dimension,
  robust
)
}
\arguments{
\item{destination}{object of class "destination"}

\item{origins}{list of objects of class "origins"}

\item{observation_time}{single integer. length of observation window in years}

\item{number_of_origins}{single integer. number of origins.}

\item{spatial_dimension}{single integer. spatial dimensions, should only be
1 or 2.}

\item{robust}{logical if TRUE robust linear model is calculated to estimate
survival and migratory connectivity}
}
\value{
object of class "mark_recapture_object": contains a list of
destinations, origins, observation time, number of origins, spatial
dimensions, empty slots for the spatial resolution, the kernel density
estimate and the estimates, the class of this object is
"mark_recapture_object"
}
\description{
This function defines the properties of the mark recapture object.
}
