% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_optimization_object}
\alias{new_optimization_object}
\title{constructor for optimization object}
\usage{
new_optimization_object(
  mark_recapture_object,
  init_beta,
  y,
  knots,
  b,
  degree,
  lambda,
  split,
  penalize,
  gradient,
  raw_spline,
  opt_beta = NULL,
  values = NULL,
  inside
)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{init_beta}{initial values for the B-spline parameters}

\item{y}{helper sequence}

\item{knots}{list of latitude and longitude of knots values for B-spline
definition}

\item{b}{character, name of breeding area to be optimized, for all breeding
areas use "all"}

\item{degree}{degree of B-spline}

\item{lambda}{numeric vector of 2, weights for the discrete and the
smoothness constraint}

\item{split}{split of the discrete non-breeding areas}

\item{penalize}{function defining penalization term for optimization}

\item{gradient}{gradient function for penalty function}

\item{raw_spline}{spline initialized over helper sequence}

\item{opt_beta}{space for the result of optimization}

\item{values}{space for the values of the optimization}

\item{inside}{logical matrix of non-breeding area specifying if a grid cell
is in the non-breeding window or not}
}
\value{
object of class "optimization_object": contains list of all the
parameters
}
\description{
This function defines the properties of an optimization object.
}
