% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_origin}
\alias{new_origin}
\title{constructor for the origin}
\usage{
new_origin(
  marked_individuals = numeric(),
  number_of_recoveries,
  migratory_connectivity
)
}
\arguments{
\item{marked_individuals}{integer: number of marked individuals in this
origin}

\item{number_of_recoveries}{numeric vector. Number of recoveries belonging to
each origin}

\item{migratory_connectivity}{function: migratory connectivity function
conditioned on this origin defined over the whole destination area}
}
\value{
object of class "origin": contains list of number of marked
individuals and migratory connectivity function
}
\description{
This function defines the properties of the origin.
}
