% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_winteringArea}
\alias{new_winteringArea}
\title{constructor for wintering area}
\usage{
new_winteringArea(window = owin(), survival, recovery, sim, data)
}
\arguments{
\item{window}{object of class "owin":observation window in
two-dimensional plane}

\item{survival}{function: survival function defined over whole
wintering area, independent of breeding area}

\item{recovery}{constant function: recovery probability, must be
constant over whole wintering area}

\item{sim}{logical. TRUE if dataset is simulated data containing true functions for
parameters. FALSE if no true parameters are available. UNDER CONSTRUCTION.}

\item{data}{logical. TRUE if dataset is real-world data. UNDER CONSTRUCTION.}
}
\value{
object of class "winteringArea": contains list of window, survival and recovery for the wintering area
}
\description{
This function defines the properties of the wintering area.
}
\examples{
new_winteringArea()
}
