% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{origin}
\alias{origin}
\title{helper function for origin}
\usage{
origin(marked_individuals, number_of_recoveries, migratory_connectivity)
}
\arguments{
\item{marked_individuals}{integer: number of marked individuals in this
origin}

\item{number_of_recoveries}{numeric vector. Number of recoveries belonging to
each origin}

\item{migratory_connectivity}{function: migratory connectivity function
conditioned on this origin defined over the whole destination area}
}
\value{
object of class "origin": contains list of number of marked
individuals and migratory connectivity function
}
\description{
This function defines the properties of the origin using the
constructor.
}
\examples{
{
 migratory_connectivity = function(b,w,B=B) {
   truncnorm::dtruncnorm(w,0,1, mean = seq(0.1,0.9,length.out = 5)[b],
   sd = 0.3)
 }
mig_con <- functional::Curry(migratory_connectivity,b=1,B=5)
b_a <- origin(marked_individuals = 10000,
    number_of_recoveries = NULL,
    migratory_connectivity = mig_con)
}
}
