% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_nf.R
\name{p_nf}
\alias{p_nf}
\title{overall probability to be not seen}
\usage{
p_nf(lb = 0, ub = 1, s, m, r, B, T)
}
\arguments{
\item{lb}{lower bound of wintering area (1D). Defaults to 0.}

\item{ub}{upper bound of wintering area (1D). Defaults to 1.}

\item{s}{function: survival function defined over whole wintering area, independent of breeding area}

\item{m}{function: migratory connectivity definded for all breeding areas and over whole wintering area}

\item{r}{constant: recovery probability, has to be constant over whole wintering area
and independent of breeding area}

\item{B}{integer, number of breeding areas}

\item{T}{integer: length of observation period}
}
\value{
list of length B: probability to be not seen independent of space and time for every breeding area
}
\description{
This function integrates the subdensity of not seen individuals.
}
