% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_grid.R
\name{par_grid}
\alias{par_grid}
\title{creates grid with function values}
\usage{
par_grid(x, y, func, ...)
}
\arguments{
\item{x}{vector of x-values}

\item{y}{vector of y-values}

\item{func}{function to create grid values}

\item{...}{optional: arguments of func}
}
\value{
length(x) x length(y) - matrix with function values evaluated at
(x,y)
}
\description{
This function creates a grid, where x are the row-names, y are the
column names and the grid-values are the function values at (x,y).
}
\examples{
par_grid(seq(0, 1, 10), seq(0, 1, 11), sum)
}
