% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_distribution.R
\name{plot_age_distribution}
\alias{plot_age_distribution}
\title{plot age distribution of raw recoveries of real world data}
\usage{
plot_age_distribution(
  mark_recapture_object,
  pdf = FALSE,
  pdf_name = "ageDistribution.pdf"
)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{pdf_name}{string to name pdf-file, please include ".pdf"!}
}
\value{
histogram of age of raw recoveries, depending on arguments plot as
pdf or to plot to device
}
\description{
This function plots the age distribution of recoveries in the real world data
as a histogram.
}
\examples{
plot_age_distribution(mro2D)

}
