% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gof_of_lm.R
\name{plot_gof_of_lm}
\alias{plot_gof_of_lm}
\title{plot goodness of fit for linear model used to estimate parameter functions}
\usage{
plot_gof_of_lm(
  mark_recapture_object,
  pdf = FALSE,
  xlb = NULL,
  xub = NULL,
  ylb = NULL,
  yub = NULL,
  draw_boundaries = TRUE,
  no_ci = FALSE,
  profile_of_parameter = NULL
)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{xlb}{numeric, lower bound of x. Defaults to NULL.}

\item{xub}{numeric, upper bound of x. Defaults to NULL.}

\item{ylb}{numeric, lower bound of y. Defaults to NULL.}

\item{yub}{numeric, upper bound of y. Defaults to NULL.}

\item{draw_boundaries}{logical, country boundaries will be drawn, if TRUE.
Defaults to TRUE.}

\item{profile_of_parameter}{sf-object containing the information of the
profile line, along which the values of a parameter can be plotted including
the bootstrap confidence interval. If this information is given, the profile
 line will be plotted}
}
\value{
matrix of dimension res*res with R^2-values for the linear model in
every point.
}
\description{
This function plots the R^2 values of the robust or ordinary linear
regression used to estimate survival.
}
\examples{
plot_gof_of_lm(mro1D)
}
