% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kde.R
\name{plot_kde}
\alias{plot_kde}
\title{plot kernel density estimate and true density for simulated data}
\usage{
plot_kde(
  b,
  mark_recapture_object,
  pdf = FALSE,
  ylim = c(0, 1.5),
  true_values_available = FALSE,
  log = FALSE,
  age_min = 0,
  age_max = NULL,
  draw_boundaries = FALSE,
  title = TRUE
)
}
\arguments{
\item{b}{specifies the area of origin for which the plot is drawn. Can be
either a origin_name, the corresponding number of the area of origin or "all"
for all areas of origin at once.}

\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{ylim}{vector in the form of c(ymin,ymax): limits of the y-axis.
Defaults to c(0,3).}

\item{true_values_available}{logical, use TRUE for simulated data, FALSE for
real-world data. Defaults to FALSE.}

\item{log}{logical, uses log-scale for kernel-density-values if TRUE.
Defaults to FALSE.}

\item{age_min}{numeric. Defaults to 0. If set only data points older than
age_min are plotted.}

\item{age_max}{numeric. If set only data points younger or equal age_max are
plotted. Defaults to NULL.}

\item{draw_boundaries}{logical, country boundaries will be drawn, if TRUE.
Defaults to TRUE.}

\item{title}{logical, if TRUE the name of the area of origin will be
added as the title. If FALSE, no title will be added. Defaults to TRUE.}
}
\value{
depending on arguments plot as pdf or to plot device
}
\description{
This function plots the kernel density estimate and true density for
simulated data.
}
\examples{
plot_kde("all", mro1D, true_values_available = TRUE)
}
