% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_m.R
\name{plot_m}
\alias{plot_m}
\title{plot true and estimated migratory connectivity}
\usage{
plot_m(
  mark_recapture_object,
  b = "all",
  pdf = FALSE,
  log = FALSE,
  true_values_available = FALSE,
  uq = 1,
  draw_boundaries = FALSE,
  no_ci = FALSE,
  profile_of_parameter = NULL
)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{b}{specifies the area of origin for which the plot is drawn. Can be
either a origin_name, the corresponding number of the areas of origin or
"all" for all areas of origin at once.}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{log}{plots logarithm of migratory connectivity. Defaults to FALSE.}

\item{true_values_available}{logical, use TRUE for simulated data, FALSE for
real-world data. Defaults to FALSE.}

\item{uq}{upper quantile until which migratory connectivity value is plotted}

\item{draw_boundaries}{logical, country boundaries will be drawn, if TRUE.
Only works if coordinates are in longitude/latitude system. Defaults to
FALSE.}

\item{no_ci}{suppresses drawing the confidence interval, even if bootstrap
information is in the mark_recapture_object. Defaults to FALSE.}

\item{profile_of_parameter}{sf-object containing the information of the
profile line, along which the values of a parameter can be plotted including
the bootstrap confidence interval. If this information is given, the profile
line will be plotted}
}
\value{
depending on arguments plot as pdf or to plot device
}
\description{
This function plots the kernel density estimate and true density for
simulated data.
}
\examples{
plot_m(mro1D, true_values_available = TRUE)
}
