% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_msr.R
\name{plot_msr}
\alias{plot_msr}
\title{plot initial situation}
\usage{
plot_msr(mark_recapture_object, pdf = FALSE, ylim = c(0, 3))
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{ylim}{vector in the form of c(ymin,ymax): limits of the y-axis.
Defaults to c(0,3).}
}
\value{
depending on arguments plot as pdf or to plot device
}
\description{
This function plots the initial functions for survival, recovery and
migratory connectivity in 1D- or 2D-space.
}
\examples{
plot_msr(mro1D)


# plot true state
}
