% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{plot parameter values (and bootstrape confidence interval, if
available) along a profile line}
\usage{
plot_profile(profile_of_parameter, ylab)
}
\arguments{
\item{profile_of_parameter}{profile of a raster object along a segmented
profile line.}

\item{ylab}{name of the y-axis, which is normally the name of the parameter}
}
\value{
vector of length res with survival probabilities dependent on space
}
\description{
This function plots the parameter values along a profile line given in the
sfc-object profile_of_parameter.
}
\examples{
{
 pop <- wrap_profile_of_param(c(0, 0), c(1, 1), mro2D, param = "s")
 plot_profile(pop, ylab = "survival")
}


}
