% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_s.R
\name{plot_s}
\alias{plot_s}
\title{plot true and estimated survival function}
\usage{
plot_s(
  mark_recapture_object,
  pdf = FALSE,
  true_values_available = FALSE,
  draw_boundaries = FALSE,
  xlb = NULL,
  zlim = c(0, 1),
  lon = NULL,
  lat = NULL,
  no_ci = FALSE,
  profile_of_parameter = NULL
)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{true_values_available}{logical, use TRUE for simulated data, FALSE for
real-world data. Defaults to FALSE.}

\item{draw_boundaries}{logical, country boundaries will be drawn, if TRUE.
Only works if coordinates are in longitude/latitude system. Defaults to
FALSE.}

\item{xlb}{if not NULL, it zooms the plot to the limits given by xlim and
ylim}

\item{zlim}{boundaries in the direction of survival values}

\item{no_ci}{suppresses drawing the confidence interval, even if bootstrap
information is in the mark_recapture_object. Defaults to FALSE.}

\item{profile_of_parameter}{sf-object containing the information of the
profile line, along which the values of a parameter can be plotted including
the bootstrap confidence interval. If this information is given, the profile
 line will be plotted}
}
\value{
vector of length res with survival probabilities dependent on space
}
\description{
This function estimates the survival from a kernel density estimate of the
data of recovered individuals. It uses the data of all areas of origin at
once.
}
\examples{
plot_s(mro1D, true_values_available = TRUE)
}
