% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgeDistribution.R
\name{plotAgeDistribution}
\alias{plotAgeDistribution}
\title{plot age distribution of raw recoveries of real world data}
\usage{
plotAgeDistribution(
  markRecaptureObject,
  pdf = FALSE,
  pdfName = "ageDistribution.pdf"
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{pdfName}{string to name pdf-file, please include ".pdf"!}
}
\value{
histogram of age of raw recoveries, depending on arguments plot as
pdf or to plot to device
}
\description{
This function plots the age distribution of recoveries in the real world data
as a histogram.
}
\examples{
plotAgeDistribution(mro2D)
}
