% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombR.R
\name{plotCombR}
\alias{plotCombR}
\title{Plot function for true and estimated continuous, discrete and combined recovery probability}
\usage{
plotCombR(
  markRecaptureObject,
  optimizationObject,
  pdf = FALSE,
  fileName = paste("rComb_", Sys.time(), ".pdf", sep = ""),
  trueContinuous = TRUE,
  trueDiscrete = TRUE,
  estContinuous = TRUE,
  estDiscrete = FALSE,
  estCombined = TRUE,
  drawBoundaries = FALSE,
  trueValuesAvailable = FALSE
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject}

\item{optimizationObject}{object of class optimizationObject}

\item{pdf}{logical, plot as pdf?, defaults to FALSE}

\item{fileName}{character string, only if pdf is TRUE, defaults to current time}

\item{trueContinuous}{logical, plot true continuous recovery probability}

\item{trueDiscrete}{logical, plot true discrete recovery probability}

\item{estContinuous}{logical, plot estimated continuous recovery probability}

\item{estDiscrete}{logical, plot estimated discrete recovery probability}

\item{estCombined}{logical, plot estimated combined recovery probability}

\item{drawBoundaries}{logical, specifies if the boundaries of the world map
should be drawn, defaults to FALSE}

\item{trueValuesAvailable}{logical, use TRUE for simulated data, FALSE for real-world data. Defaults to FALSE.}
}
\description{
This function plots (in 1D) combinations of true and estimated recovery probability values. It is possible to chose between continuous, discrete and combined estimates.
}
\examples{
{
    oO <- optimizationObject(markRecaptureObject = mro1DIncreasing$mro,
        b = "all",
        split = mro1DIncreasing$split,
        lambda  = c(.05,300))

    plotCombR(mro1DIncreasing$mro,oO)
}
}
