% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFsr.R
\name{plotFsr1D}
\alias{plotFsr1D}
\title{plot initial situation in 1D-space}
\usage{
plotFsr1D(markRecaptureObject, pdf = FALSE, ylim = c(0, 3))
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{ylim}{vector in the form of c(ymin,ymax): limits of the y-axis. Defaults to c(0,3).}
}
\value{
depending on arguments plot as pdf or to plot device
}
\description{
This function plots the initial functions for survival, recovery and
migratory connectivity in 1D-space.
}
\examples{
plotFsr1D()
}
