% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKDE.R
\name{plotKDE1D}
\alias{plotKDE1D}
\title{plot kernel density estimate and true density for simulated data}
\usage{
plotKDE1D(b, kde, res_x, markRecaptureObject, pdf = FALSE, ylim = c(0,
  1.5))
}
\arguments{
\item{b}{index of breeding area}

\item{kde}{kernel density estimate from raw data}

\item{res_x}{resolution in x direction}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{ylim}{vector in the form of c(ymin,ymax): limits of the y-axis. Defaults to c(0,3).}
}
\value{
depending on arguments plot as pdf or to plot device
}
\description{
This function plots the kernel density estimate and true density for simulated data.
}
\examples{
plotFsr1D()
}
