% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_param.R
\name{plotly_param}
\alias{plotly_param}
\title{plot 3D plot using plotly}
\usage{
plotly_param(
  mark_recapture_object,
  param,
  b = "all",
  filename = "",
  no_ci = FALSE
)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{param}{parameter to be plotted. Can be "m" or "s". Only one parameter
can be plotted at once. Select different areas of origin with the argument
b.}

\item{b}{specifies the area of origin for which the plot is drawn. Can be
either a origin_name, the corresponding number of the area of origin or "all"
for all origins at once.}

\item{filename}{optional filename extension. Defaults to "".}

\item{no_ci}{suppresses drawing the confidence interval, even if bootstrap
information is in the mark_recapture_object. Defaults to FALSE.}
}
\value{
saves a html file and returns the plotly plot
}
\description{
This function plots the kernel density estimate and true density for
simulated data.
}
\examples{
{
mro2 <- est_uncertainty(mro2D, "s", iterations = 2, filename = "test")
mro2 <- bootstrap_quantiles(mro2, "s", "test")
plotly_param(mro2, param = "s")
}

}
