% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyParam.R
\name{plotlyParam}
\alias{plotlyParam}
\title{plot 3D plot using plotly}
\usage{
plotlyParam(markRecaptureObject, param, b = "all", filename = "", noCI = FALSE)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{param}{parameter to be plotted. Can be "m" or "s". Only one parameter
can be plotted at once. Select different areas of origin with the argument
b.}

\item{b}{specifies breeding area for which the plot is drawn. Can be either
a breedingAreaName, the corresponding number of the breeding area or "all"
for all breeding areas at once.}

\item{filename}{optional filename extension. Defaults to "".}

\item{noCI}{suppresses drawing the confidence interval, even if bootstrap
information is in the markRecaptureObject. Defaults to FALSE.}
}
\value{
saves a html file and returns the plotly plot
}
\description{
This function plots the kernel density estimate and true density for
simulated data.
}
\examples{
plotM(mro1D, trueValuesAvailable = TRUE)
}
