% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotM.R
\name{plotM}
\alias{plotM}
\title{plot true and estimated migratory connectivity}
\usage{
plotM(res_x, markRecaptureObject, all = FALSE, pdf = FALSE,
  log = FALSE, dataType = "sim")
}
\arguments{
\item{res_x}{resolution in x direction}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{m_fit}{estimate for m, matrix}
}
\value{
depending on arguments plot as pdf or to plot device
}
\description{
This function plots the kernel density estimate and true density for simulated data.
}
\examples{
plotM()

}
