% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{plotProfile}
\alias{plotProfile}
\title{Defaults to TRUE.
plot parameter values (and bootstrape confidence interval, if
available) along a profile line}
\usage{
plotProfile(profileOfParameter, ylab)
}
\arguments{
\item{profileOfParameter}{profile of a raster object along a segmented
profile line.}

\item{ylab}{name of the y-axis, which is normally the name of the parameter}
}
\value{
vector of length res with survival probabilities dependent on space
}
\description{
This function plots the parameter values along a profile line given in the
sfc-object profileOfParameter.
}
\examples{
plotProfile(profileOfParameter, ylab = "survival")
}
