% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotS.R
\name{plotS}
\alias{plotS}
\title{plot true and estimated survival function}
\usage{
plotS(
  res,
  markRecaptureObject,
  pdf = FALSE,
  dataType = "sim",
  xlim = NULL,
  ylim = NULL,
  drawBoundaries = TRUE,
  xlb = NULL,
  zlim = c(0, 1)
)
}
\arguments{
\item{res}{numeric, spatial resolution for longitude and latitude}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{dataType}{character, use "sim" for simulated data, "data" for real world data. Defaults to "sim".}

\item{xlim}{vector of lower bound and upper bound of x. Defaults to NULL.}

\item{ylim}{vector of lower bound and upper bound of y. Defaults to NULL.}

\item{drawBoundaries}{logical, country boundaries will be drawn, if TRUE. Defaults to TRUE.}

\item{xlb}{if not NULL, it zooms the plot to the limits given by xlim and ylim}

\item{zlim}{boundaries in the direction of survival values}
}
\value{
vector of length res with survival probabilities dependent on space
}
\description{
This function estimates the survival from a kernel density estimate of the data
of recovered individuals. It uses the data of all breeding areas at once.
}
\examples{
plotS()
}
