% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_line.R
\name{profile_line}
\alias{profile_line}
\title{Creates a linear profile line.}
\usage{
profile_line(coord_start, coord_end, crs)
}
\arguments{
\item{coord_start}{coordinates of the starting point of the line as
longitude/latitude. E.g., c(0,0).}

\item{coord_end}{coordinates of the end point of the line as
longitude/latitude. E.g., c(0,0).}

\item{crs}{coordinate system of mark_recapture_object}
}
\value{
profile line as simple feature object.
}
\description{
This function converts the start and end points from longitude/latitude
to utm coordinates and creates a continuous line between these points as a
simple feature object.
}
\examples{
profile_line(c(0, 0), c(1, 1), crs = "ESRI:54009")

}
