% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_points.R
\name{profile_points}
\alias{profile_points}
\title{Parameter values along a segmentized profile line}
\usage{
profile_points(profile_line, df_max_length = 60000)
}
\arguments{
\item{profile_line}{coordinates of the starting point of the line as
longitude/latitude. E.g., c(0,0).}

\item{df_max_length}{parameter of the function sf::st_segmentize.}
}
\value{
sfc-object including the profile line as points and the distance
 on the profile line
}
\description{
This function segmentizes a continous line (sf-object) into points
to calculates the distance between these points
}
\examples{
profile_points(profile_line(c(0, 0), c(1, 1), crs = "ESRI:54009"))

}
