% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileLine.R
\name{profileLine}
\alias{profileLine}
\title{Creates a linear profile line.}
\usage{
profileLine(coordStartPoint, coordEndPoint)
}
\arguments{
\item{coordStartPoint}{coordinates of the starting point of the line as
longitude/latitude. E.g., c(0,0).}

\item{coordEndPoint}{coordinates of the end point of the line as
longitude/latitude. E.g., c(0,0).}
}
\value{
profile line as simple feature object.
}
\description{
This function converts the start and end points from longitude/latitude
to utm coordinates and creates a continuous line between these points as a
simple feature object.
}
\examples{
profileLine(c(0,0),c(1,1))
}
