% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileOfParameter.R
\name{profileOfParameter}
\alias{profileOfParameter}
\title{Parameter values along a profile line}
\usage{
profileOfParameter(rasteredParam, profilePoints)
}
\arguments{
\item{rasteredParam}{raster stack of the parameter values and the
bootstrap confidence interval}

\item{profilePoints}{sfc-object including the profile line as points and
the distance on the profile line}

\item{param}{name of the parameter}
}
\value{
values of the given parameter along a profile line
}
\description{
This function calculates the profile of a raster object along a segmented
profile line.
}
\examples{
profileOfParameter(rasterParam(mro2D, "s", "all"),
profilePoints(profileLine(c(0,0),c(1,1))))
}
