% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_window.R
\name{project_window}
\alias{project_window}
\title{projects owin-object to another coordinate system}
\usage{
project_window(win, old_crs = "EPSG:4326", new_crs = "ESRI:54009")
}
\arguments{
\item{win}{owin-object}

\item{old_crs}{coordinate system of the input owin-object. Defaults to
longitude/latitude "EPSG:4326"}

\item{new_crs}{coordinate system of the output owin-object. Defaults to the
Mollweide Projection "ESRI:54009"}
}
\value{
owin-object with projected coordinate-system
}
\description{
This function projects an owin object to another coordinate system, which
can be specified.
}
\examples{
new_win <- project_window(mro2D$destination$win,
old_crs = mro2D$destination$crs)

}
