% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectDf.R
\name{projectDf}
\alias{projectDf}
\title{projects data.frame to another coordinate system}
\usage{
projectDf(
  df,
  lon = "longitude",
  lat = "latitude",
  old_crs = "EPSG:4326",
  new_crs = "ESRI:54009"
)
}
\arguments{
\item{df}{data.frame with columns for coordinates, e.g., longitude and
latitude}

\item{lon}{column name of the column containing information on longitude}

\item{lat}{column name of the column containing information on latitude}

\item{old_crs}{coordinate system of the input owin-object. Defaults to
longitude/latitude "EPSG:4326"}

\item{new_crs}{coordinate system of the output owin-object. Defaults to the
Mollweide Projection "ESRI:54009"}
}
\value{
data.frame with projected coordinate-system
}
\description{
This function projects a data.frame to another coordinate system, which
can be specified.
}
\examples{
new_df <- projectDf(mro2D$winteringArea$recoveryData$b1)
}
