% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectWindow.R
\name{projectWindow}
\alias{projectWindow}
\title{projects owin-object to another coordinate system}
\usage{
projectWindow(win, old_crs = "EPSG:4326", new_crs = "ESRI:54009")
}
\arguments{
\item{win}{owin-object}

\item{old_crs}{coordinate system of the input owin-object. Defaults to
longitude/latitude "EPSG:4326"}

\item{new_crs}{coordinate system of the output owin-object. Defaults to the
Mollweide Projection "ESRI:54009"}
}
\value{
owin-object with projected coordinate-system
}
\description{
This function projects an owin object to another coordinate system, which
can be specified.
}
\examples{
new_win <- projectWindow(mro2D$winteringArea$win)
}
