% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_param.R
\name{raster_param}
\alias{raster_param}
\title{Raster stack of parameter values and bootstrap confidence intervals, if
available.}
\usage{
raster_param(mark_recapture_object, param, b)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{param}{name of the parameter}

\item{b}{name of area of origin}
}
\value{
raster stack of parameter value and bootstrap confidence interval.
}
\description{
This function creates a raster stack of the parameter value and the 2.5%-
and the 97.5%-quantile of the bootstrapped data.
}
\examples{
raster_param(mro2D, "s", "all")

}
