% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterParam.R
\name{rasterParam}
\alias{rasterParam}
\title{Raster stack of parameter values and bootstrap confidence intervals, if
available.}
\usage{
rasterParam(markRecaptureObject, param, b)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{param}{name of the parameter}

\item{b}{name of area of origin}
}
\value{
raster stack of parameter value and bootstrap confidence interval.
}
\description{
This function creates a raster stack of the parameter value and the 2.5%-
and the 97.5%-quantile of the bootstrapped data.
}
\examples{
rasterParam(mro2D, "s", "all")
}
