% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_inds_func.R
\name{rec_inds_func}
\alias{rec_inds_func}
\title{calculates number of recovered individuals from marked and recoveries data
frame}
\usage{
rec_inds_func(origin_names, recoveries)
}
\arguments{
\item{origin_names}{character vector with names of origins}

\item{recoveries}{data frame of recoveries with column on marking location}
}
\value{
vector of number of recovered individuals per area of origin
}
\description{
This function calculates the number of recovered individuals marked in one
area of origin given the data frame of dead recoveries and a vector of names
of areas of origin
}
\examples{
rec_inds_func(c("early", "late"), robinsInit$recovery_data)
}
