% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recIndsFunc.R
\name{recIndsFunc}
\alias{recIndsFunc}
\title{calculates number of recovered individuals from marked and recoveries data frame}
\usage{
recIndsFunc(breedingAreaNames, recoveries)
}
\arguments{
\item{breedingAreaNames}{character vector with names of breeding areas}

\item{recoveries}{data frame of recoveries with column on marking location}
}
\value{
vector of number of recovered individuals per breeding area
}
\description{
This function calculates the number of recovered individuals marked in one breeding area
given the data frame of dead recoveries and a vector of names of breeding areas
}
\examples{
recIndsFunc(c("early","late"),robinsInit$recoveryData)
}
