% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robins.R
\docType{data}
\name{robins}
\alias{robins}
\title{example for real-world data: robins}
\format{
A markRecaptureObject.
\describe{
  \item{robins}{markRecaptureObject}
}
}
\usage{
data(robins)
}
\description{
A dataset containing the raw data of robins marked at the ringing station
Greifswalder Oie, Germany, during autumn migration and dead recoveries between
November and February in Europe and Africa. Additionally, kernel
density, survival, migratory connectivity and recovery probability are estimated
by the continuous approach and stored in the markRecaptureObject.
}
\examples{

data(robins)
str(robins)

}
\references{
The data is provided by the Verein Jordsand e.V. and the Beringungszentrale Hiddensee.
}
\keyword{datasets}
