% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_contin.R
\name{sim_contin}
\alias{sim_contin}
\title{simulation function for mark-recovery setting}
\usage{
sim_contin(mark_recapture_object, res = 100)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}
}
\value{
returns object of class mark_recapture_object with added simulated
recovery data
}
\description{
This function allows you to simulate data of dead recoveries with specified
survival, migratory connectivity and recovery probability in 1D- or 2D-space
and time using rejection sampling.
}
\examples{
{
  mro <- mark_recapture_object(
    xrange = c(0, 1),
    survival = function(w) {
      0.5 * w + .4
    },
    recovery = function(w) {
      0.01
    },
    marked_individuals = rep(1000, 3),
    migratory_connectivity = function(b, w, B = 3) {
      truncnorm::dtruncnorm(w, 0, 1,
        mean = seq(0.1, 0.9, length.out = B)[b],
        sd = 0.3
      )
    },
    observationTime = 5
  )
  mro <- sim_contin(mro)
}

}
