% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simContin.R
\name{simContin}
\alias{simContin}
\title{simulation function for mark-recovery setting}
\usage{
simContin(markRecaptureObject)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}
}
\value{
returns object of class markRecaptureObject with added simulated
recoveryData
}
\description{
This function allows you to simulate data of dead recoveries with specified
survival, migratory connectivity and recovery probability in 1D- or 2D-space
and time using rejection sampling.
}
\examples{
{
 mro <- markRecaptureObject(xrange = c(0,1),
                            survival = function(w){0.5*w+.4},
                            recovery = function(w){0.01},
                            markedInds = rep(1000,3),
                            migratoryConnectivity = function(b,w,B=3){
                              truncnorm::dtruncnorm(w,0,1,
                              mean = seq(0.1,0.9,length.out = B)[b],
                              sd = 0.3)},
                           observationTime = 5)
mro <- simContin(mro)
}
}
