% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_profile_of_param.R
\name{wrap_profile_of_param}
\alias{wrap_profile_of_param}
\title{Parameter values along a profile line}
\usage{
wrap_profile_of_param(
  coord_start,
  coord_end,
  mark_recapture_object,
  param,
  b = "all"
)
}
\arguments{
\item{coord_start}{coordinates of the starting point of the line as
longitude/latitude. E.g., c(0,0).}

\item{coord_end}{coordinates of the end point of the line as
longitude/latitude. E.g., c(0,0).}

\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{param}{name of the parameter}

\item{b}{name of origin. Defaults to "all".}
}
\value{
values of the given parameter along a profile line
}
\description{
This function creates a linear profile line between two points,
segmentizes it into points and calculates the distance between the segments.
It is a wrapper function around all necessary steps. To customize this
procedure, use the functions inside.
}
\examples{
wrap_profile_of_param(c(0, 0), c(1, 1), mro2D, "s")

}
