% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapProfileOfParam.R
\name{wrapProfileOfParam}
\alias{wrapProfileOfParam}
\title{Parameter values along a profile line}
\usage{
wrapProfileOfParam(
  coordStartPoint,
  coordEndPoint,
  markRecaptureObject,
  param,
  b = "all"
)
}
\arguments{
\item{coordStartPoint}{coordinates of the starting point of the line as
longitude/latitude. E.g., c(0,0).}

\item{coordEndPoint}{coordinates of the end point of the line as
longitude/latitude. E.g., c(0,0).}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{param}{name of the parameter}
}
\value{
values of the given parameter along a profile line
}
\description{
This function creates a linear profile line between two points,
segmentizes it into points and calculates the distance between the segments.
It is a wrapper function around all necessary steps. To customize this
procedure, use the functions inside.
}
\examples{
wrapProfileOfParam(c(0,0),c(1,1), mro2D, "s")
}
