% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_quantiles.R
\name{bootstrap_quantiles}
\alias{bootstrap_quantiles}
\title{linear model of kernel density estimate}
\usage{
bootstrap_quantiles(mark_recapture_object, param = c("s", "m", "r"), filename)
}
\arguments{
\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}

\item{filename}{(path to file) and filename to store the
mark_recapture_object with bootstrap estimates.}

\item{parameters}{vector of characters specifying the parameter names
considered for uncertainty estimation, possible values are s for survival,
m for migratory connectivity and r for recovery probability. Defaults to
c("s","m","r").}
}
\value{
raw bootstrap dataframe to be used in bootstrap_quantiles
}
\description{
This function estimates the uncertainty of the parameters survival, migratory
connectivity and recovery probability by bootstrapping the marking data.
}
\examples{
mro <- bootstrap_quantiles(mro1D, "s")

}
