% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapMarkingData.R
\name{bootstrapMarkingData}
\alias{bootstrapMarkingData}
\title{linear model of kernel density estimate}
\usage{
bootstrapMarkingData(markRecaptureObject)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}
}
\value{
markRecaptureObject with estimates of all selected parameters
}
\description{
This function bootstraps the marking data for every area of origin.
}
\examples{
mro <- bootstrapMarkingData(mro1D)
}
