% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{breedingArea}
\alias{breedingArea}
\title{helper function for breeding area}
\usage{
breedingArea(markedInds, numberOfRecoveries, migratoryConnectivity)
}
\arguments{
\item{markedInds}{integer: number of marked individuals in this breeding area}

\item{numberOfRecoveries}{numeric vector. Number of recoveries belonging to each breeding area.}

\item{migratoryConnectivity}{function: migratory connectivity function conditioned
on this breeding area defined over whole wintering area}
}
\value{
object of class "breedingArea": contains list of number of marked individuals
and migratory connectivity function
}
\description{
This function defines the properties of the breeding area using the constructor.
}
\examples{
{
migratoryConnectivity = function(b,w,B=B){
    truncnorm::dtruncnorm(w,0,1, mean = seq(0.1,0.9,length.out = 5)[b], sd = 0.3)
    }
migCon <- functional::Curry(migratoryConnectivity,b=1,B=5)
bA <- breedingArea(markedInds = 10000,
    numberOfRecoveries = NULL,
    migratoryConnectivity = migCon)
}
}
