% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{breedingArea}
\alias{breedingArea}
\title{helper function for breeding area}
\usage{
breedingArea(markedInds, numberOfRecoveries, migratoryConnectivity,
  sim = NULL, data = NULL)
}
\arguments{
\item{markedInds}{integer: number of marked individuals in this breeding area}

\item{migratoryConnectivity}{function: migratory connectivity function conditioned
on this breeding area defined over whole wintering area}
}
\value{
object of class "breedingArea": contains list of number of marked individuals
and migratory connectivity function
}
\description{
This function defines the properties of the breeding area using the constructor.
}
\examples{
breedingArea()
}
