% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estKDE.R
\name{estKDE}
\alias{estKDE}
\title{estimate raw density of recovered individuals}
\usage{
estKDE(
  markRecaptureObject,
  res,
  all = FALSE,
  dataType = "sim",
  xname = "x",
  yname = "y",
  timename = "time",
  bw = NULL
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{res}{resolution in space}

\item{all}{boolean: if TRUE only one kernel density estimate will be calculated
summarising all breeding areas. Defaults to FALSE.}

\item{dataType}{character. Specifies the type of data used. Possible values are "sim" for
simulated data by the simContin-function and "data" for external data. Defaults to "sim".}

\item{xname}{name of x variable, e.g. longitude}

\item{yname}{name of y variable, e.g. latitude}

\item{timename}{name of time variable, e.g. age}

\item{bw}{numeric. bandwidth parameter. Defaults to NULL.}
}
\value{
markRecaptureObject with list of values created by sparr::spattemp.density (see ?sparr::spattemp.density for details) and spatial resolution.
}
\description{
This function estimates the raw density of recovered individuals in space and time
from given data using a kernel density estimate.
}
\examples{
estKDE()
}
