% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estKDE.R
\name{estKDE}
\alias{estKDE}
\title{estimate raw density of recovered individuals}
\usage{
estKDE(markRecaptureObject, res_x, all = FALSE, dataType = "sim",
  xname = "x", yname = "y", timename = "time", h = NULL)
}
\arguments{
\item{res_x}{resolution in space}

\item{all}{boolean: if TRUE only one kernel density estimate will be calculated
summarising all breeding areas. Defaults to FALSE.}

\item{xname}{name of x variable, e.g. longitude}

\item{yname}{name of y variable, e.g. latitude}

\item{timename}{name of time variable, e.g. age}

\item{eta}{list with B entries: data of recovered individuals.
eta[[b]][,1] contains the 1-dimensional space of recovery,
eta[[b]][,2] contains the integer age at recovery}

\item{B}{integer, number of breeding areas}

\item{T}{integer: length of observation period}
}
\value{
list of values created by sparr::spattemp.density (see ?sparr::spattemp.density for details)
}
\description{
This function estimates the raw density of recovered individuals in space and time
from given data using a kernel density estimate.
}
\examples{
estKDE()
}
