% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estLM.R
\name{estLM}
\alias{estLM}
\title{linear model of kernel density estimate}
\usage{
estLM(
  res,
  markRecaptureObject,
  b,
  dataType = "sim",
  robust = T,
  auxiliaryVariable = NULL
)
}
\arguments{
\item{res}{spatial resolution for longitude and latitude}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{dataType}{type of data. Simulated data: "sim", reals world data: "data".
Defaults to "sim".}

\item{robust}{type of estimator used for linear regression to estimate survival. If TRUE
a robust regression will be computed (robustbase::lmrob()), if FALSE an ordinary regression
(base::lm()). Defaults to TRUE.}

\item{auxiliaryVariable}{specify, if an auxiliary variable like the at risk-population should be
used. Defaults to NULL.}
}
\value{
vector of length res with survival probabilities dependent on space
}
\description{
This function estimates the parameters of a linear model for a certain point
of the kernel density estimate dependent on time. Linear models are obtained
according to the resolution.
}
\examples{
estLM()
}
