% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estLM.R
\name{estLM}
\alias{estLM}
\title{linear model of kernel density estimate}
\usage{
estLM(markRecaptureObject, b, fixedSlope = NULL, auxiliaryVariable = NULL)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{fixedSlope}{numeric. Value for the fixed slope, e.g., to estimate
a linear model for the breeding areas separately.}

\item{auxiliaryVariable}{specify, if an auxiliary variable like the at risk-population should be
used. Defaults to NULL.}
}
\value{
vector of length res with survival probabilities dependent on space
}
\description{
This function estimates the parameters of a linear model for a certain point
of the kernel density estimate dependent on time. Linear models are obtained
according to the resolution.
}
\examples{
estLM()
}
