% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estM.R
\name{estM}
\alias{estM}
\title{estimate migratory connectivity}
\usage{
estM(markRecaptureObject, all = FALSE, auxiliaryVariable = NULL)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{all}{boolean: if TRUE only one kernel density estimate will be calculated
summarising all breeding areas. Defaults to FALSE.}

\item{auxiliaryVariable}{specify, if an auxiliary variable like the at risk-population should be
used. Defaults to NULL.#' @return list of vectors with length res-1 containing migratory connectivity density of every spot}
}
\description{
This function estimates the migratory connectivity in space constant over time
when survival and raw distribution of dead recoveries for every breeding area is known.
}
\examples{
mro <- estM(mro1D, all = TRUE)
}
