% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estM.R
\name{estM}
\alias{estM}
\title{estimate migratory connectivity}
\usage{
estM(res_x, markRecaptureObject, s_fit, all = FALSE)
}
\arguments{
\item{res_x}{resolution in space}

\item{all}{boolean: if TRUE only one kernel density estimate will be calculated
summarising all breeding areas. Defaults to FALSE.}

\item{B}{integer, number of breeding areas}

\item{kde}{list of B values created by sparr::spattemp.density
(see ?sparr::spattemp.density for details): raw distribution of dead recoveries}
}
\value{
list of vectors with length res_x-1 containing migratory connectivity density of every spot
}
\description{
This function estimates the migratory connectivity in space constant over time
when survival and raw distribution of dead recoveries for every breeding area is known.
}
\examples{
estF()
}
