% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estS.R
\name{estS}
\alias{estS}
\title{survival estimator}
\usage{
estS(res_x, kde_all, T)
}
\arguments{
\item{res_x}{resolution in space}

\item{kde_all}{list of one value created by sparr::spattemp.density
(see ?sparr::spattemp.density for details): raw distribution of dead recoveries}

\item{T}{integer: length of observation period}
}
\value{
vector of length res_x with survival probabilities dependent on space
}
\description{
This function estimates the survival from a kernel density estimate of the data
of recovered individuals. It uses the data of all breeding areas at once.
}
\examples{
estS()
}
