% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estS.R
\name{estS}
\alias{estS}
\title{survival estimator}
\usage{
estS(res_x, markRecaptureObject, res_y = res_x, dataType = "sim",
  robust = T)
}
\arguments{
\item{res_x}{resolution in space}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}
}
\value{
vector/matrix of length res_x with survival probabilities dependent on space
}
\description{
This function estimates the survival from a kernel density estimate of the data
of recovered individuals. It uses the data of all breeding areas at once.
}
\examples{
estS()
}
