% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estUncertainty.R
\name{estUncertainty}
\alias{estUncertainty}
\title{linear model of kernel density estimate}
\usage{
estUncertainty(
  markRecaptureObject,
  parameters = c("s", "m", "r"),
  iterations = 100
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{parameters}{vector of characters specifying the parameter names
considered for uncertainty estimation, possible values are s for survival,
m for migratory connectivity and r for recovery probability. Defaults to
c("s","m","r").}

\item{iterations}{numeric. Number of bootstraps performed. Defaults to 100.}
}
\value{
markRecaptureObject with added bootstrap uncertainty for parameters
}
\description{
This function estimates the uncertainty of the parameters survival, migratory
connectivity and recovery probability by bootstrapping the marking data.
}
\examples{
mro <- estUncertainty(mro1D)
}
