% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_f.R
\name{f_f}
\alias{f_f}
\title{density function for recovered individuals only}
\usage{
f_f(w, t, b, markRecaptureObject, p)
}
\arguments{
\item{w}{decimal number: spatial point of recovery (1-dimensional)}

\item{t}{integer: temporal component of recovery}

\item{b}{integer: breeding area}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{p}{desimal: complementary probability to be not found}
}
\value{
density of recovered individuals for the specified parameters
}
\description{
This function sets up a valid density function for every spatiotemporal point for recovered individuals
by means of given survival, migratory connectivity and recovery probability.
}
\examples{
f_f()
}
