% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_f.R
\name{f_f}
\alias{f_f}
\title{density function for recovered individuals only}
\usage{
f_f(w, t, b, lb, ub, s, m, r, T)
}
\arguments{
\item{w}{decimal number: spatial point of recovery (1-dimensional)}

\item{t}{integer: temporal component of recovery}

\item{b}{integer: breeding area}

\item{lb}{decimal: lower bound of wintering area}

\item{ub}{decimal: upper bound of wintering area}

\item{s}{function: survival function defined over whole wintering area, independent of breeding area}

\item{m}{function: migratory connectivity definded for all breeding areas and over whole wintering area}

\item{r}{constant: recovery probability, has to be constant over whole wintering area
and independent of breeding area}

\item{T}{integer: length of observation period}

\item{f_f}{function: subdensity of recovered individuals}

\item{f_nf}{function: subdensity of not recovered individuals}
}
\value{
density of recovered individuals for the specified parameters
}
\description{
This function sets up a valid density function for every spatiotemporal point for recovered individuals
by means of given survival, migratory connectivity and recovery probability.
}
\examples{
f_f()
}
