% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_f.R
\name{f_f}
\alias{f_f}
\title{density function for recovered individuals only}
\usage{
f_f(w, t, b, markRecaptureObject, p)
}
\arguments{
\item{w}{decimal number (1D) or vector of decimal numbers (2D): spatial point
of recovery}

\item{t}{integer: temporal component of recovery}

\item{b}{specifies breeding area for which the plot is drawn. Can be either
a breedingAreaName, the corresponding number of the breeding area or "all"
for all breeding areas at once.}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{p}{decimal: complementary probability to be not found}
}
\value{
density of recovered individuals for the specified parameters
}
\description{
This function sets up a valid density function for every spatio-temporal
point for recovered individuals by means of given survival, migratory
connectivity and recovery probability.
}
\examples{
{
    p <- 1-p_nf(b=1,mro1D)
    ff <- f_f(1,1,1,mro1D,p)
}
}
