% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_f_sub.R
\name{f_f_sub}
\alias{f_f_sub}
\title{subdensity function for recovered individuals}
\usage{
f_f_sub(w, t, b, markRecaptureObject)
}
\arguments{
\item{w}{decimal number: spatial point of recovery (1-dimensional)}

\item{t}{integer: temporal component of recovery}

\item{b}{integer: breeding area}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}
}
\value{
subdensity of recovered individuals for the specified parameters
}
\description{
This function sets up the subdensity function for every spatiotemporal point for recovered individuals
by means of given survival, migratory connectivity and recovery probability.
}
\examples{
f_f_sub()
}
