% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.R
\name{gr}
\alias{gr}
\title{gradient of penalizing function for combined parameter estimate}
\usage{
gr(beta, rawSpline, m, b, lambda, split, A, prop, dim, res, inside, normalize)
}
\arguments{
\item{beta}{vector of parameters of the optimization function}

\item{m}{vector of continuous migratory connectivity estimates}

\item{lambda}{weights for different penalization terms}

\item{split}{vector of length of y which defines the affiliation to a discrete
wintering area}

\item{prop}{vector of proportions of individuals going to discrete wintering areas
(discrete estimate or expected value for migratory connectivity)}

\item{inside}{specifies if a cell of the gridded window is inside the window of the data
or not. Vector of logicals.}

\item{y}{auxiliary vector spanning the whole wintering area}

\item{knots}{vector of inner knots for b-spline definition}

\item{degree}{degree of bspline}
}
\value{
function depending on bspline parameters, which returns the sum of
        quadratic distances to continuous and discrete migratory connectivity
        and smoothness
}
\description{
This function defines the penalization term for the combined estimation approach.
It can be used to jointly optimize distance to continuous migratory connectivity estimates
gained by estM, discrete migratory connectivity estimates and maximize smoothness.
}
