% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateDist2Discrete.R
\name{integrateDist2Discrete}
\alias{integrateDist2Discrete}
\title{function to define the distance between a b-spline representing a density
and the discrete proportion}
\usage{
integrateDist2Discrete(
  rawSpline,
  dim,
  split,
  beta,
  b,
  res,
  prop,
  print = TRUE,
  inside,
  normalize
)
}
\arguments{
\item{rawSpline}{raw spline, value of initSpline()}

\item{dim}{spatial dimension of the data}

\item{split}{vector of length of y which defines the affiliation to a discrete
wintering area}

\item{beta}{parameter of the b-spline, values are to be optimized}

\item{b}{name of breeding area}

\item{res}{spatial resolution of the estimates}

\item{prop}{vector of proportions of individuals going to discrete wintering areas
(discrete estimate or expected value for migratory connectivity)}

\item{print}{logical, should proportions be printed or not?}

\item{inside}{specifies if a cell of the gridded window is inside the window of the data
or not. Vector of logicals.}
}
\value{
function defining the distance between a b-spline and
        discrete migratory connectivity depending on the parameters
}
\description{
This function numerically integrates the quadratic distance between a b-spline
representing a density and the discrete migratory connectivity
}
\examples{
integrateDist2Discrete()
}
