% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh.R
\name{lh}
\alias{lh}
\title{function defining the smoothness of a bspline}
\usage{
lh(beta, dim, A, normalize)
}
\arguments{
\item{beta}{vector of parameters of the b-spline}

\item{dim}{dimension of space}

\item{A}{matrix of second derivative of b-spline}

\item{normalize}{factor to normalize with.}
}
\value{
function defining the distance between a b-spline and
        discrete migratory connectivity depending on the parameters
}
\description{
This function defines the smoothness of a bspline as the second derivative
}
\examples{
{
    y <- seq(0,100222.75,length.out=100)
    i_k <- seq(0, 100222.75, length.out = 10)[2:9]
    A <- splines2::dbs(y, knots=i_k, derivs = 2, degree = 3,
                        intercept = TRUE)
    L <- lh(beta, dim = 1,
        A = t(A) \%*\% A,
        normalize = 100 / 100222.75^4)
    L(beta = rnorm(12))
}
}
