% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lh.R
\name{Lh}
\alias{Lh}
\title{function defining the smoothness of a bspline}
\usage{
Lh(beta, dim, A, res, normalize)
}
\arguments{
\item{beta}{vector of parameters of the b-spline}

\item{dim}{dimension of space}

\item{A}{matrix of second derivative of b-spline}

\item{res}{spatial resolution of the estimates}

\item{inside}{specifies if a cell of the gridded window is inside the window of the data
or not. Vector of logicals.}
}
\value{
function defining the distance between a b-spline and
        discrete migratory connectivity depending on the parameters
}
\description{
This function defines the smoothness of a bspline as the second derivative
}
\examples{
Lh(w,beta,knots,degree)
}
