% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_breedingArea}
\alias{new_breedingArea}
\title{constructor for breeding area}
\usage{
new_breedingArea(markedInds = numeric(), numberOfRecoveries,
  migratoryConnectivity, sim, data)
}
\arguments{
\item{markedInds}{integer: number of marked individuals in this breeding area}

\item{migratoryConnectivity}{function: migratory connectivity function conditioned
on this breeding area defined over whole wintering area}
}
\value{
object of class "breedingArea": contains list of number of marked individuals
and migratory connectivity function
}
\description{
This function defines the properties of the breeding area.
}
\examples{
new_breedingArea()
}
