% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_markRecaptureObject}
\alias{new_markRecaptureObject}
\title{constructor for mark recapture object}
\usage{
new_markRecaptureObject(
  winteringArea,
  breedingAreas,
  observationTime,
  numberOfBreedingAreas,
  spatialDim,
  robust
)
}
\arguments{
\item{winteringArea}{object of class "winteringArea"}

\item{breedingAreas}{list of objects of class "breedingAreas"}

\item{observationTime}{single integer. length of observation window in years}

\item{numberOfBreedingAreas}{single integer. number of breeding areas.}

\item{spatialDim}{single integer. spatial dimensions, should only be 1 or 2.}

\item{robust}{logical if TRUE robust linear model is calculated to estimate survival and migratory connectivity}
}
\value{
object of class "markRecaptureObject": contains list of wintering area,
breeding areas, observationTime, number of breeding areas, spatial dimensions,
empty slots for the spatial resolution, the kernel density estimate and the estimates,
the class of this object is "markRecaptureObject"
}
\description{
This function defines the properties of the mark recapture object.
}
\examples{
new_markRecaptureObject()

}
