% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_nf.R
\name{p_nf}
\alias{p_nf}
\title{overall probability to be not seen}
\usage{
p_nf(b, mark_recapture_object)
}
\arguments{
\item{b}{specifies the area of origin for which the plot is drawn. Can be
either a origin_name, the corresponding number of the area of origin or "all"
for all areas of origin at once.}

\item{mark_recapture_object}{object of class mark_recapture_object
(see mark_recapture_object())}
}
\value{
list: probability to be not seen independent of space
and time for every area of origin
}
\description{
This function integrates the subdensity of not seen individuals.
}
\examples{
p_nf(b = "b1", mro1D)

}
