% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_nf.R
\name{p_nf}
\alias{p_nf}
\title{overall probability to be not seen}
\usage{
p_nf(b, markRecaptureObject)
}
\arguments{
\item{b}{specifies breeding area for which the plot is drawn. Can be either a breedingAreaName,
the corresponding number of the breeding area or "all" for all breeding areas at once.}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}
}
\value{
list: probability to be not seen independent of space
and time for every breeding area
}
\description{
This function integrates the subdensity of not seen individuals.
}
\examples{
p_nf(b = "b1", mro1D)
}
