% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombM.R
\name{plotCombM}
\alias{plotCombM}
\title{Plot function for true and estimated continuous, discrete and combined migratory connectivity}
\usage{
plotCombM(
  markRecaptureObject,
  optimizationObject,
  pdf = FALSE,
  fileName = paste("mComb_", Sys.time(), ".pdf", sep = ""),
  trueContinuous = TRUE,
  trueDiscrete = TRUE,
  estContinuous = TRUE,
  estDiscrete = FALSE,
  estCombined = TRUE,
  estCorrected = FALSE,
  zlim = NULL,
  drawBoundaries = FALSE,
  trueValuesAvailable = FALSE,
  uq = 1
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject}

\item{optimizationObject}{object of class optimizationObject}

\item{pdf}{logical, plot as pdf?, defaults to FALSE}

\item{fileName}{character string, only if pdf is TRUE, defaults to current time}

\item{trueContinuous}{logical, plot true continuous migratory connectivity}

\item{trueDiscrete}{logical, plot true discrete migratory connectivity}

\item{estContinuous}{logical, plot estimated continuous migratory connectivity}

\item{estDiscrete}{logical, plot estimated discrete migratory connectivity}

\item{estCombined}{logical, plot estimated combined migratory connectivity}

\item{trueValuesAvailable}{logical, use TRUE for simulated data, FALSE for real-world data. Defaults to FALSE.}
}
\description{
This function plots (in 1D) combinations of true and estimated migratory connectivity values. It is possible to chose between continuous, discrete and combined estimates.
}
\examples{
plotCombM()
}
