% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGOFofLM.R
\name{plotGOFofLM}
\alias{plotGOFofLM}
\title{plot goodness of fit for linear model used to estimate parameter functions}
\usage{
plotGOFofLM(
  markRecaptureObject,
  pdf = FALSE,
  xlb = NULL,
  xub = NULL,
  ylb = NULL,
  yub = NULL,
  drawBoundaries = TRUE,
  noCI = FALSE
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{xlb}{numeric, lower bound of x. Defaults to NULL.}

\item{xub}{numeric, upper bound of x. Defaults to NULL.}

\item{ylb}{numeric, lower bound of y. Defaults to NULL.}

\item{yub}{numeric, upper bound of y. Defaults to NULL.}

\item{drawBoundaries}{logical, country boundaries will be drawn, if TRUE.
Defaults to TRUE.}
}
\value{
matrix of dimension res*res with R^2-values for the linear model in
every point.
}
\description{
This function plots the R^2 values of the robust or ordinary linear
regression used to estimate survival.
}
\examples{
plotGOFofLM(mro1D)
}
