% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGOFofLM.R
\name{plotGOFofLM}
\alias{plotGOFofLM}
\title{plot goodness of fit for linear model used to estimate parameter functions}
\usage{
plotGOFofLM(
  res_x,
  res_y = res_x,
  markRecaptureObject,
  pdf = FALSE,
  dataType = "sim",
  xlb = NULL,
  xub = NULL,
  ylb = NULL,
  yub = NULL,
  drawBoundaries = TRUE
)
}
\arguments{
\item{res_x}{resolution in space}

\item{res_y}{numeric. resolution in space in north-south direction. Defaults to res_x.}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{dataType}{character, use "sim" for simulated data, "data" for real world data. Defaults to "sim".}

\item{drawBoundaries}{logical, country boundaries will be drawn, if TRUE. Defaults to TRUE.}

\item{xlim}{vector of lower bound and upper bound of x. Defaults to NULL.}

\item{ylim}{vector of lower bound and upper bound of y. Defaults to NULL.}
}
\value{
matrix of dimension res_x*res_y with R^2-values for the linear model in every point.
}
\description{
This function plots the R^2 values of the robust or ordinary linear regression
used to estimate survival.
}
\examples{
plotGOFofS()
}
