% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKDE.R
\name{plotKDE}
\alias{plotKDE}
\title{plot kernel density estimate and true density for simulated data}
\usage{
plotKDE(
  b,
  res,
  markRecaptureObject,
  pdf = FALSE,
  ylim = c(0, 1.5),
  dataType = "sim",
  log = FALSE,
  facetByTime = TRUE,
  drawBoundaries = TRUE
)
}
\arguments{
\item{b}{specifies breeding area for which the plot is drawn. Can be either a breedingAreaName,
the corresponding number of the breeding area or "all" for all breeding areas at once.}

\item{res}{resolution for longitude and latitude}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{ylim}{vector in the form of c(ymin,ymax): limits of the y-axis. Defaults to c(0,3).}

\item{dataType}{character, use "sim" for simulated data, "data" for real world data. Defaults to "sim".}

\item{log}{logical, uses log-scale for kernel-density-values if TRUE. Defaults to FALSE.}

\item{facetByTime}{logical, plots one plot for every age class if TRUE. Defaults to TRUE.}

\item{drawBoundaries}{logical, country boundaries will be drawn, if TRUE. Defaults to TRUE.}
}
\value{
depending on arguments plot as pdf or to plot device
}
\description{
This function plots the kernel density estimate and true density for simulated data.
}
\examples{
plotKDE()
}
