% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawRecoveries.R
\name{plotRawRecoveries}
\alias{plotRawRecoveries}
\title{plot raw recoveries of real world data}
\usage{
plotRawRecoveries(markRecaptureObject, pdf = FALSE,
  pdfName = "rawRecoveries.pdf", areaNames = NULL)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{pdfName}{string to name pdf-file, please include ".pdf"!}

\item{areaNames}{optional, order of breeding area names to be plotted, defaults to NULL.
In the default case order of breeding areas in markRecaptureObject will be chosen.}
}
\value{
depending on arguments plot as pdf or to plot to device
}
\description{
This function plots the distribution of recoveries in the real world data in Europe and Africa.
}
\examples{
plotRawRecoveries()
}
