% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawRecoveries.R
\name{plotRawRecoveries}
\alias{plotRawRecoveries}
\title{plot raw recoveries of real world data}
\usage{
plotRawRecoveries(
  markRecaptureObject,
  pdf = FALSE,
  pdfName = "rawRecoveries.pdf",
  areaNames = NULL,
  facetByAge = FALSE,
  facetByArea = FALSE,
  ageMin = 0,
  ageMax = NULL,
  xname = "longitude",
  yname = "latitude",
  timename = "age",
  markAreaName = "markArea",
  plotTitle = ""
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{pdfName}{string to name pdf-file, please include ".pdf"!}

\item{areaNames}{optional, order of breeding area names to be plotted, defaults to NULL.
In the default case order of breeding areas in markRecaptureObject will be chosen.}

\item{facetByAge}{logical, determines if the data should be facetted by the age column.
Defaults to FALSE.}

\item{facetByArea}{logical, determines if the data should be facetted by the area column.
Defaults to FALSE.}

\item{ageMin}{numeric. Defaults to 0. If set only data points older than ageMin are plotted.}

\item{ageMax}{numeric. If set only data points younger or equal ageMax are plotted. Defaults to NULL.}

\item{left}{numeric. Western border of map. Defaults to -24.}

\item{bottom}{numeric. Southern border of map. Defaults to -35.}

\item{right}{numeric. Eastern border of map. Defaults to 71.}

\item{top}{numeric. Northern border of map. Defaults to 71.}
}
\value{
depending on arguments plot as pdf or to plot to device
}
\description{
This function plots the distribution of recoveries in the real world data in Europe and Africa.
}
\examples{
plotRawRecoveries()
}
