% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawRecoveries.R
\name{plotRawRecoveries}
\alias{plotRawRecoveries}
\title{plot raw recoveries of real world data}
\usage{
plotRawRecoveries(
  markRecaptureObject,
  pdf = FALSE,
  pdfName = "rawRecoveries.pdf",
  areaNames = NULL,
  facetByAge = FALSE,
  facetByArea = FALSE,
  ageMin = 0,
  ageMax = NULL,
  xname = "longitude",
  yname = "latitude",
  timename = "age",
  markAreaName = "markArea",
  plotTitle = "",
  map = NULL,
  prj = "+proj=longlat",
  transformToPrj = "+proj=longlat"
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{pdfName}{string to name pdf-file, please include ".pdf"!}

\item{areaNames}{optional, order of breeding area names to be plotted,
defaults to NULL.
In the default case order of breeding areas in markRecaptureObject will be
chosen.}

\item{facetByAge}{logical, determines if the data should be faceted by the
age column. Defaults to FALSE.}

\item{facetByArea}{logical, determines if the data should be faceted by
the area column.
Defaults to FALSE.}

\item{ageMin}{numeric. Defaults to 0. If set only data points older than
ageMin are plotted.}

\item{ageMax}{numeric. If set only data points younger or equal ageMax are
plotted. Defaults to NULL.}

\item{xname}{character, name of column containing first dimension, defaults
to "longitude".}

\item{yname}{character, name of column containing second dimension, defaults
to "latitude".}

\item{timename}{character, name of column containing third dimension,
defaults to "age".}

\item{markAreaName}{character, name of column containing information on
areas where marking took place, defaults to "markArea".}

\item{plotTitle}{character, title of plot, defaults to "".}

\item{map}{map to plot the data on. Needs to be compatible with ggplot2.
For example, map can be ggmap::ggmap(myMap), where myMap is
created by ggmap::get_stamenmap.}

\item{prj}{projection of the coordinates of the dead recoveries. Defaults to
"+proj=longlat". For details see ?sf::st_crs. If the data has another
projection it will be transformed to the longitude latitude system
using sf::st_transform}

\item{transformToPrj}{projection the dead recoveries are transformed to.
Defaults to "+proj=longlat". Should be the projection of the map.
Data will only be transformed when transformToPrj does not equal prj.}
}
\value{
depending on arguments plot as pdf or to plot to device
}
\description{
This function plots the distribution of recoveries in the real world data in
Europe and Africa.
}
\examples{
plotRawRecoveries(mro1D)
}
