% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotS.R
\name{plotS}
\alias{plotS}
\title{plot true and estimated survival function}
\usage{
plotS(res_x, markRecaptureObject, pdf = FALSE)
}
\arguments{
\item{res_x}{resolution in space}

\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}
}
\value{
vector of length res_x with survival probabilities dependent on space
}
\description{
This function estimates the survival from a kernel density estimate of the data
of recovered individuals. It uses the data of all breeding areas at once.
}
\examples{
plotS()
}
