% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotS.R
\name{plotS}
\alias{plotS}
\title{Defaults to TRUE.
plot true and estimated survival function}
\usage{
plotS(
  markRecaptureObject,
  pdf = FALSE,
  trueValuesAvailable = FALSE,
  drawBoundaries = FALSE,
  xlb = NULL,
  zlim = c(0, 1),
  longitude = NULL,
  latitude = NULL,
  noCI = FALSE,
  profileOfParameter = NULL
)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}

\item{pdf}{logical, saves image as pdf-file if TRUE. Defaults to FALSE.}

\item{trueValuesAvailable}{logical, use TRUE for simulated data, FALSE for
real-world data. Defaults to FALSE.}

\item{drawBoundaries}{logical, country boundaries will be drawn, if TRUE.
Only works if coordinates are in longitude/latitude system. Defaults to FALSE.}

\item{xlb}{if not NULL, it zooms the plot to the limits given by xlim and
ylim}

\item{zlim}{boundaries in the direction of survival values}

\item{noCI}{suppresses drawing the confidence interval, even if bootstrap
information is in the markRecaptureObject. Defaults to FALSE.}

\item{profileOfParameter}{sf-object containing the information of the profile
line, along which the values of a parameter can be plotted including the
 bootstrap confidence interval. If this information is given, the profile
 line will be plotted}
}
\value{
vector of length res with survival probabilities dependent on space
}
\description{
This function estimates the survival from a kernel density estimate of the
data of recovered individuals. It uses the data of all breeding areas at
once.
}
\examples{
plotS(mro1D, trueValuesAvailable = TRUE)
}
