% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_of_parameter.R
\name{profile_of_parameter}
\alias{profile_of_parameter}
\title{Parameter values along a profile line}
\usage{
profile_of_parameter(rastered_param, profile_points)
}
\arguments{
\item{rastered_param}{raster stack of the parameter values and the
bootstrap confidence interval}

\item{profile_points}{sfc-object including the profile line as points and
the distance on the profile line}

\item{param}{name of the parameter}
}
\value{
values of the given parameter along a profile line
}
\description{
This function calculates the profile of a raster object along a segmented
profile line.
}
\examples{
profile_of_parameter(
  rasterParam(mro2D, "s", "all"),
  profile_points(profile_line(c(0, 0), c(1, 1)))
)

}
