% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_mark_recapture.R
\name{project_mark_recapture}
\alias{project_mark_recapture}
\title{projects mark_recapture_project to another coordinate system}
\usage{
project_mark_recapture(
  mark_recapture_project,
  lon = "longitude",
  lat = "latitude",
  old_crs = "EPSG:4326",
  new_crs = "ESRI:54009"
)
}
\arguments{
\item{lon}{column name of the column containing information on longitude}

\item{lat}{column name of the column containing information on latitude}

\item{old_crs}{coordinate system of the input owin-object. Defaults to
longitude/latitude "EPSG:4326"}

\item{new_crs}{coordinate system of the output owin-object. Defaults to the
Mollweide Projection "ESRI:54009"}

\item{df}{data.frame with columns for coordinates, e.g., longitude and
latitude}
}
\value{
data.frame with projected coordinate-system
}
\description{
This function projects all relevant components of a mark_recapture_project to
another coordinate system, which can be specified. These components are
the recovery data and the window.
}
\examples{
mro2 <- project_mark_recapture(mro2D)

}
