% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recIndsFunc.R
\name{recIndsFunc}
\alias{recIndsFunc}
\title{calculates number of recovered individuals from marked and recoveries data frame}
\usage{
recIndsFunc(marked, recoveries)
}
\arguments{
\item{marked}{data frame with column for marking location names}

\item{recoveries}{data frame of recoveries with column on marking location}
}
\value{
vector of number of recovered individuals per breeding area
}
\description{
This function calculates the number of recovered individuals marked in one breeding area
given the data frame of dead recoveries and optionally a
}
\examples{
recIndsFunc()
}
