% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simContin.R
\name{simContin}
\alias{simContin}
\title{1D simulation function}
\usage{
simContin(B, p_nf, f_f, n)
}
\arguments{
\item{B}{integer: total number of breeding areas}

\item{p_nf}{vector with length B: probability of not seen
individuals from breeding area b independent of place of recovery (marginalized over whole wintering area)}

\item{f_f}{function: density of found individuals only}

\item{n}{vector of length B: number of marked individuals per breeding area}
}
\value{
list of vector k: number of recovered dividuals and list eta with B entries, every entry containing
a 2xk - data.frame of space and time of every recovery
}
\description{
This function allows you to simulate data of dead recoveries with specified
survival, migratory connectivity and recovery probability in 1D-space and time
using rejection sampling.
}
\examples{
simContin()
}
