% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simContin.R
\name{simContin}
\alias{simContin}
\title{simulation function for mark-recovery setting}
\usage{
simContin(markRecaptureObject)
}
\arguments{
\item{markRecaptureObject}{object of class markRecaptureObject
(see markRecaptureObject())}
}
\value{
returns object of class markRecaptureObject with added simulated recoveryData:
list of vector k: number of recovered dividuals and list eta with B entries, every entry containing
a 2xk - data.frame of space and time of every recovery
}
\description{
This function allows you to simulate data of dead recoveries with specified
survival, migratory connectivity and recovery probability in 1D- or 2D-space and time
using rejection sampling.
}
\examples{
simContin()
}
