% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{winteringArea}
\alias{winteringArea}
\title{helper function for wintering area}
\usage{
winteringArea(
  window = NULL,
  survival,
  recovery,
  xrange = c(0, 0),
  yrange = c(0, 0),
  recoveryData = NULL
)
}
\arguments{
\item{window}{object of class "owin":observation window in
two-dimensional plane}

\item{survival}{function: survival function defined over whole
wintering area, independent of breeding area}

\item{recovery}{constant function: recovery probability, must be
constant over whole wintering area}

\item{xrange}{vector in the form of c(xmin,xmax). To define line or rectangle xrange and yrange can be used instead of window.}

\item{yrange}{vector in the form of c(ymin,ymax). To define line or rectangle xrange and yrange can be used instead of window.}

\item{recoveryData}{empty space for simulated or real-world recovery data}
}
\value{
object of class "winteringArea": contains list of window, survival and recovery for the wintering area
}
\description{
This function defines the properties of the wintering area using the constructor.
}
\examples{
winteringArea()
}
